:recycle: refactor(generator-engine): split 3384-line god file into focused sub-modules
Extract each generator domain into apps/web/src/lib/services/seo/generators/:
- base.ts: nameTable, nameGeneratorConfig, GeneratorOutput, pickFrom, getRandomItems, generateName
- npc.ts: npcConfig, npcThemeConfig, generateNPC
- faction.ts: factionConfig, vampireConfig, themeIdToLabel, generateFaction, generateVampireClan
- settlement.ts: settlementConfig, generateSettlement
- magic-item.ts: magicItemConfig, generateMagicItem
- quest.ts: questConfig, generateQuestHook
- names.ts: generateNames
- social-hub.ts: socialHubConfig, generateSocialHub, generateTavern
- kingdom-nation.ts: nationConfig, kingdomConfig, generateKingdom, generateNation

generator-engine.ts reduced to a 75-line barrel that re-exports all configs and wraps
standalone functions in the DefaultGeneratorEngine facade. All 1877 tests pass.
