+++ apps/web/src/lib/services/seo/generator-engine.ts
@@ -1,3383 +1,107 @@
 import { aiClientManager } from "$lib/services/ai/client-manager";
 
-// Local Name Tables
-export const nameTable = {
-  prefixes: [
-    "Ael",
-    "Bran",
-    "Cael",
-    "Dax",
-    "El",
-    "Fael",
-    "Grom",
-    "Had",
-    "Ior",
-    "Jor",
-    "Kael",
-    "Lyr",
-    "Morg",
-    "Nal",
-    "Odh",
-    "Py",
-    "Quel",
-    "Rhu",
-    "Sari",
-    "Thor",
-    "Uth",
-    "Vael",
-    "Wulf",
-    "Xan",
-    "Ysv",
-    "Zeph",
-  ],
-  suffixes: [
-    "dar",
-    "gorn",
-    "thor",
-    "ius",
-    "eth",
-    "wen",
-    "dolf",
-    "ric",
-    "mar",
-    "mon",
-    "las",
-    "vian",
-    "dor",
-    "morn",
-    "ra",
-    "thas",
-    "val",
-    "gar",
-    "rin",
-    "kis",
-    "on",
-    "ar",
-    "en",
-    "is",
-    "a",
-    "os",
-  ],
-  descriptors: [
-    "The Brave",
-    "The Silent",
-    "The Swift",
-    "The Bold",
-    "The Seer",
-    "The Wise",
-    "The Shadow",
-    "The Iron",
-    "The Wild",
-    "The Grim",
-    "The Lost",
-    "The Seeker",
-    "The Wanderer",
-    "The Exile",
-    "The Fierce",
-    "The Chronicler",
-  ],
-};
+export {
+  nameTable,
+  nameGeneratorConfig,
+  type GeneratorOutput,
+  pickFrom,
+  getRandomItems,
+  generateName,
+} from "./generators/base";
+export { npcConfig, npcThemeConfig } from "./generators/npc";
+export {
+  factionConfig,
+  themeIdToLabel,
+  vampireConfig,
+} from "./generators/faction";
+export { settlementConfig } from "./generators/settlement";
+export { magicItemConfig } from "./generators/magic-item";
+export { questConfig } from "./generators/quest";
+export { socialHubConfig } from "./generators/social-hub";
+export { nationConfig, kingdomConfig } from "./generators/kingdom-nation";
 
-// Fantasy Name Generator Config
-export const nameGeneratorConfig = {
-  cultures: [
-    "Generic Fantasy",
-    "High Elf",
-    "Dark Elf",
-    "Dwarven",
-    "Halfling",
-    "Orcish",
-    "Norse / Viking",
-    "Roman / Latin",
-    "Celtic / Gaelic",
-    "Eastern / Asian-inspired",
-  ],
-  genders: ["Any", "Masculine", "Feminine", "Neutral / Ambiguous"],
-  nameTypes: ["Person", "Place", "Faction", "Item"],
-  counts: ["3", "5", "10"],
-  // Culture-specific local fallback tables
-  culturePrefixes: {
-    "Generic Fantasy": [
-      "Ael",
-      "Bran",
-      "Cael",
-      "Dax",
-      "El",
-      "Fael",
-      "Kael",
-      "Lyr",
-      "Morg",
-      "Nal",
-      "Thor",
-      "Vael",
-    ],
-    "High Elf": [
-      "Aer",
-      "Caer",
-      "El",
-      "Gal",
-      "Ith",
-      "Lae",
-      "Mir",
-      "Sil",
-      "Thal",
-      "Var",
-      "Wyn",
-      "Zin",
-    ],
-    "Dark Elf": [
-      "Driz",
-      "Mal",
-      "Nath",
-      "Riz",
-      "Sol",
-      "Szor",
-      "Ul",
-      "Vir",
-      "Xan",
-      "Zaer",
-      "Zin",
-      "Vy",
-    ],
-    Dwarven: [
-      "Bal",
-      "Dur",
-      "Glor",
-      "Grim",
-      "Kat",
-      "Kaz",
-      "Mag",
-      "Nar",
-      "Thor",
-      "Tor",
-      "Ulf",
-      "Vor",
-    ],
-    Halfling: [
-      "Bil",
-      "Cor",
-      "Del",
-      "Fil",
-      "Gob",
-      "Mer",
-      "Pip",
-      "Rose",
-      "Sam",
-      "Tom",
-      "Wil",
-      "Bun",
-    ],
-    Orcish: [
-      "Druk",
-      "Grak",
-      "Grom",
-      "Krag",
-      "Mog",
-      "Nar",
-      "Rok",
-      "Skul",
-      "Thok",
-      "Urg",
-      "Vrak",
-      "Zug",
-    ],
-    "Norse / Viking": [
-      "Arn",
-      "Bjor",
-      "Dag",
-      "Eil",
-      "Gunn",
-      "Hal",
-      "Ing",
-      "Leif",
-      "Rag",
-      "Sigr",
-      "Thor",
-      "Ulf",
-    ],
-    "Roman / Latin": [
-      "Aem",
-      "Aur",
-      "Bru",
-      "Cal",
-      "Cas",
-      "Cor",
-      "Flav",
-      "Jul",
-      "Marc",
-      "Octa",
-      "Sext",
-      "Val",
-    ],
-    "Celtic / Gaelic": [
-      "Aed",
-      "Bran",
-      "Caill",
-      "Conn",
-      "Donn",
-      "Eoch",
-      "Fearg",
-      "Niall",
-      "Ruad",
-      "Taig",
-      "Uar",
-      "Cunn",
-    ],
-    "Eastern / Asian-inspired": [
-      "Akira",
-      "Chen",
-      "Hiro",
-      "Jin",
-      "Kai",
-      "Li",
-      "Ren",
-      "Ryu",
-      "Shen",
-      "Tao",
-      "Wei",
-      "Zhen",
-    ],
-  } as Record<string, string[]>,
-  cultureSuffixes: {
-    "Generic Fantasy": [
-      "dar",
-      "eth",
-      "gorn",
-      "ius",
-      "mar",
-      "morn",
-      "ra",
-      "ric",
-      "thas",
-      "val",
-      "wen",
-      "thor",
-    ],
-    "High Elf": [
-      "ael",
-      "aith",
-      "ara",
-      "ath",
-      "elan",
-      "iath",
-      "ien",
-      "ira",
-      "is",
-      "ith",
-      "on",
-      "rial",
-    ],
-    "Dark Elf": [
-      "ace",
-      "arn",
-      "bra",
-      "dra",
-      "fein",
-      "ica",
-      "inae",
-      "ra",
-      "rix",
-      "ssa",
-      "tra",
-      "vra",
-    ],
-    Dwarven: [
-      "ak",
-      "ald",
-      "ar",
-      "ek",
-      "im",
-      "in",
-      "kin",
-      "nar",
-      "rek",
-      "rim",
-      "rin",
-      "uk",
-    ],
-    Halfling: [
-      "ble",
-      "boro",
-      "buck",
-      "fast",
-      "foot",
-      "fur",
-      "good",
-      "hill",
-      "penny",
-      "seed",
-      "wise",
-      "wool",
-    ],
-    Orcish: [
-      "ash",
-      "dar",
-      "gash",
-      "grak",
-      "gruk",
-      "kul",
-      "lash",
-      "mok",
-      "rak",
-      "skrag",
-      "thak",
-      "ugh",
-    ],
-    "Norse / Viking": [
-      "ald",
-      "bjorn",
-      "dís",
-      "eit",
-      "gar",
-      "helm",
-      "nar",
-      "rún",
-      "stein",
-      "ulf",
-      "vald",
-      "var",
-    ],
-    "Roman / Latin": [
-      "anus",
-      "ella",
-      "ia",
-      "ianus",
-      "illa",
-      "ina",
-      "inus",
-      "io",
-      "ius",
-      "lia",
-      "ona",
-      "us",
-    ],
-    "Celtic / Gaelic": [
-      "ach",
-      "all",
-      "an",
-      "ard",
-      "as",
-      "dha",
-      "ech",
-      "en",
-      "inn",
-      "on",
-      "ua",
-      "ugh",
-    ],
-    "Eastern / Asian-inspired": [
-      "bo",
-      "fa",
-      "ji",
-      "ko",
-      "lan",
-      "lei",
-      "li",
-      "mei",
-      "na",
-      "ro",
-      "shi",
-      "xia",
-    ],
-  } as Record<string, string[]>,
-};
-
-// NPC Generator Table Config
-export const npcConfig = {
-  races: [
-    "Human",
-    "Elf",
-    "Dwarf",
-    "Halfling",
-    "Orc",
-    "Gnome",
-    "Tiefling",
-    "Dragonborn",
-  ],
-  roles: [
-    "Mage",
-    "Warrior",
-    "Rogue",
-    "Priest",
-    "Merchant",
-    "Scholar",
-    "Blacksmith",
-    "Guard",
-    "Noble",
-    "Innkeeper",
-  ],
-  alignments: [
-    "Lawful Good",
-    "Neutral Good",
-    "Chaotic Good",
-    "Lawful Neutral",
-    "True Neutral",
-    "Chaotic Neutral",
-    "Lawful Evil",
-    "Neutral Evil",
-    "Chaotic Evil",
-  ],
-  traits: [
-    "Always whispers when speaking to build dramatic tension.",
-    "Carries a pocket watch that runs backward but claims it is correct.",
-    "Extremely superstitious about black cats and wooden doors.",
-    "Has a collection of rare, dried flowers in their cloak pockets.",
-    "Never looks anyone directly in the eye, shifting their gaze constantly.",
-    "Speaks in rhyming riddles when they become nervous or excited.",
-    "Has a nervous twitch in their left hand when speaking about magic.",
-    "Obsessed with cleanliness, frequently wiping down their gear.",
-  ],
-  secrets: [
-    "Is a secret spy for a rival merchant guild operating in the shadows.",
-    "Possesses a cursed map that shows their exact death location, which is nearby.",
-    "Accidentally poisoned their previous master and fled the crime scene.",
-    "Is actually a shapechanger in hiding from an ancient wizard.",
-    "Stole a sacred relic from the local temple and keeps it in their boot.",
-    "Is deeply in debt to a dangerous local crime lord who wants them dead.",
-    "Knows the secret passcode to the royal vault under the capital.",
-  ],
-  motives: [
-    "To clear their family's disgraced name and reclaim their ancestral land.",
-    "To fund a search for a lost sibling who vanished in the Underdark.",
-    "To acquire enough wealth to buy their freedom from a contract.",
-    "To find a cure for a mysterious family affliction affecting their bloodline.",
-    "To exact revenge on the corrupt noble who exiled them.",
-    "To locate a legendary magical spellbook hidden in a nearby ruin.",
-  ],
-  factions: [
-    "The Ashen Ledger, a quiet network of debt collectors and informants.",
-    "The Lantern Court, a civic order that keeps public peace after sunset.",
-    "The Red Sash Company, sellswords with a habit of choosing winning sides.",
-    "The Argent Loom, an artisan guild that hides coded messages in its work.",
-    "The Chapel of Last Mercy, a temple faction that knows too many confessions.",
-    "The Blackwater Compact, smugglers moving relics beneath legitimate trade.",
-  ],
-  plotHooks: [
-    "They ask the party to recover a sealed letter before it reaches a rival.",
-    "They recognize one character from a prophecy but refuse to explain in public.",
-    "They can open a locked district gate if the party solves their immediate problem.",
-    "They are being followed by someone who disappears whenever challenged.",
-    "They offer a reward for escort, then reveal the destination is forbidden ground.",
-    "They own a clue that reframes a recent villain as someone else's pawn.",
-  ],
-};
-
-// Settlement Generator Table Config
-export const settlementConfig = {
-  sizes: [
-    { name: "Hamlet", range: "50-100 inhabitants", pointsOfInterestCount: 1 },
-    { name: "Village", range: "100-500 inhabitants", pointsOfInterestCount: 2 },
-    { name: "Town", range: "500-5000 inhabitants", pointsOfInterestCount: 3 },
-    { name: "City", range: "5000-20000 inhabitants", pointsOfInterestCount: 4 },
-  ],
-  economies: [
-    "Agriculture",
-    "Mining",
-    "Trade Hub",
-    "Fishing",
-    "Black Market",
-    "Arcane Study",
-  ],
-  governments: [
-    "Council of Elders",
-    "Feudal Lord",
-    "Merchant Oligarchy",
-    "Military Dictatorship",
-    "Democracy",
-    "Magocracy",
-  ],
-  notableLocations: [
-    "The Rusty Anchor Tavern",
-    "Temple of the Sun",
-    "Grand Archive",
-    "Whispering Woods Gate",
-    "Vault of Secrets",
-    "Alchemist's Greenhouse",
-    "Market Bazaar",
-    "Ruined Watchtower",
-  ],
-  factions: [
-    "The Iron Shield Guard",
-    "The Shadow Thieves Guild",
-    "The Whispering Monks",
-    "The Gilded Merchants",
-    "The Arcane Assembly",
-  ],
-};
-
-// Magic Item Table Config
-export const magicItemConfig = {
-  types: [
-    "Weapon",
-    "Armor",
-    "Wand",
-    "Ring",
-    "Amulet",
-    "Potion",
-    "Scroll",
-    "Wondrous Item",
-  ],
-  rarities: ["Common", "Uncommon", "Rare", "Very Rare", "Legendary"],
-  properties: [
-    "Glows faintly in the presence of undead, flashing red when danger is close.",
-    "Whispers ancient prophecies to the wearer in their sleep, which are usually vague.",
-    "Allows the user to speak with small forest animals, but they only talk about food.",
-    "Increases the user's running speed, but leaves a trail of harmless sparks.",
-    "Can store a single memory to be retrieved later by touching the surface.",
-    "Grants resistance to cold, but the wearer always feels slightly chilly inside.",
-  ],
-  histories: [
-    "Forged in the heart of a dying star by ancient dwarven smiths of old.",
-    "Recovered from the hoard of a red dragon that ravaged the northern kingdoms.",
-    "Worn by a legendary paladin who fell during the Siege of Shadowkeep.",
-    "Discovered inside a hollow tree trunk deep within the Feywild.",
-    "Created by a mad wizard who vanished into their own pocket dimension.",
-  ],
-};
-
-// Faction Generator Table Config
-export const factionConfig = {
-  themes: [
-    "Classic Fantasy",
-    "Cyberpunk / Corporate",
-    "Vampire / Gothic Noir",
-    "Sci-Fi / Space Opera",
-    "Modern Conspiracy",
-    "Post-Apocalyptic",
-  ],
-  types: [
-    "Merchant Guild",
-    "Secret Society",
-    "Mercenary Company",
-    "Temple Order",
-    "Criminal Syndicate",
-    "Rebel Cell",
-    "Arcane Circle",
-  ],
-  scopes: [
-    "Local district",
-    "Single city",
-    "Border region",
-    "Trade route",
-    "Hidden stronghold",
-    "Kingdom-wide network",
-  ],
-  alignments: [
-    "Publicly lawful, privately ruthless",
-    "Idealistic but compromised",
-    "Pragmatic and profit-driven",
-    "Fanatical and secretive",
-    "Protective of common folk",
-    "Opportunistic and divided",
-  ],
-  goals: [
-    "Control a contested trade route before a rival power does.",
-    "Recover a forbidden relic buried beneath a civic landmark.",
-    "Replace corrupt officials with loyal agents.",
-    "Protect a hidden sanctuary from outside discovery.",
-    "Break an old treaty that limits their expansion.",
-    "Expose a rival faction's crimes without revealing their own.",
-  ],
-  conflicts: [
-    "A splinter leader is selling secrets to an enemy.",
-    "Their public mission conflicts with the methods they use at night.",
-    "A recent victory created debts they cannot repay.",
-    "Their patron has vanished, leaving rival lieutenants in charge.",
-    "A hostage, ledger, or relic could unravel their legitimacy.",
-    "Their members disagree over whether the party is useful or dangerous.",
-  ],
-  hooks: [
-    "They hire the party for a simple delivery that is actually a loyalty test.",
-    "They ask for protection during a meeting with a bitter rival.",
-    "They offer information about a villain in exchange for public help.",
-    "They frame the party to force them into negotiation.",
-    "They need outsiders to enter a place their members are forbidden to visit.",
-    "They ask the party to choose between two bad successors.",
-  ],
-};
-
-export const themeIdToLabel: Record<string, string> = {
-  fantasy: "Classic Fantasy",
-  fantasy_dark: "Classic Fantasy",
-  cyberpunk: "Cyberpunk / Corporate",
-  cyberpunk_light: "Cyberpunk / Corporate",
-  horror: "Vampire / Gothic Noir",
-  horror_light: "Vampire / Gothic Noir",
-  scifi: "Sci-Fi / Space Opera",
-  scifi_light: "Sci-Fi / Space Opera",
-  modern: "Modern Conspiracy",
-  modern_dark: "Modern Conspiracy",
-  apocalyptic: "Post-Apocalyptic",
-  apocalyptic_light: "Post-Apocalyptic",
-};
-
-export const npcThemeConfig = {
-  ancestries: {
-    "Classic Fantasy": [
-      "Human",
-      "Elf",
-      "Dwarf",
-      "Halfling",
-      "Tiefling",
-      "Half-Orc",
-      "Gnome",
-      "Dragonborn",
-    ],
-    "Cyberpunk / Corporate": [
-      "Human",
-      "Street-Modified Human",
-      "Corporate Clone",
-      "Synthetic Android",
-      "Uplifted Organism",
-    ],
-    "Vampire / Gothic Noir": [
-      "Human",
-      "Dhampir",
-      "Revenant",
-      "Changed Mortal",
-      "Witchblood",
-    ],
-    "Sci-Fi / Space Opera": [
-      "Human",
-      "Android",
-      "Colony-Born",
-      "Alien Citizen",
-      "Augmented Pilot",
-    ],
-    "Modern Conspiracy": [
-      "Human",
-      "Off-Grid Survivor",
-      "Enhanced Operative",
-      "Whistleblower",
-    ],
-    "Post-Apocalyptic": [
-      "Survivor Human",
-      "Mutant",
-      "Scavenger-Born",
-      "Vault Dweller",
-      "Wasteland Nomad",
-    ],
-  } as Record<string, string[]>,
-  roles: {
-    "Classic Fantasy": [
-      "Mage",
-      "Warrior",
-      "Rogue",
-      "Priest",
-      "Merchant",
-      "Noble",
-      "Scholar",
-      "Guard",
-    ],
-    "Cyberpunk / Corporate": [
-      "Netrunner",
-      "Street Fixer",
-      "Corporate Agent",
-      "Street Samurai",
-      "Techie",
-      "Gang Lieutenant",
-      "Medtech",
-      "Journalist",
-    ],
-    "Vampire / Gothic Noir": [
-      "Vampire Hunter",
-      "Occultist",
-      "Corrupt Noble",
-      "Private Detective",
-      "Fallen Clergy",
-      "Criminal Enforcer",
-      "Asylum Keeper",
-    ],
-    "Sci-Fi / Space Opera": [
-      "Starship Pilot",
-      "Engineer",
-      "Colonial Marine",
-      "Diplomat",
-      "Free Trader",
-      "Scientist",
-      "AI Liaison",
-    ],
-    "Modern Conspiracy": [
-      "Intelligence Agent",
-      "Investigative Journalist",
-      "Fixer",
-      "Activist",
-      "Corporate Operative",
-      "Private Investigator",
-    ],
-    "Post-Apocalyptic": [
-      "Scavenger",
-      "Wasteland Warlord",
-      "Medic",
-      "Trader",
-      "Cult Enforcer",
-      "Scout",
-      "Mechanic",
-    ],
-  } as Record<string, string[]>,
-  moralities: {
-    "Classic Fantasy": [
-      {
-        id: "chivalric_code",
-        label: "Chivalric Code",
-        aiPromptDirective:
-          "Write this NPC with an unshakeable sense of honor, classical righteousness, and duty. Their vocabulary is formal, respectful, and rejects deceitful or underhanded methods.",
-      },
-      {
-        id: "common_good",
-        label: "Common Good",
-        aiPromptDirective:
-          "This NPC is driven entirely by empathy and the immediate welfare of the community. They are casual, caring, and willing to quietly break unjust rules to protect others.",
-      },
-      {
-        id: "enlightened_balance",
-        label: "Enlightened Balance",
-        aiPromptDirective:
-          "Write this NPC as an objective, emotionally level philosopher who prioritizes cosmic or natural balance. Avoid emotional outbursts or blind loyalty to any faction.",
-      },
-      {
-        id: "mercenary_instinct",
-        label: "Mercenary Instinct",
-        aiPromptDirective:
-          "This NPC is intensely pragmatic, apolitical, and motivated by coins, safety, or trade. Their dialogue should be transactional, street-smart, and grounded in cold reality.",
-      },
-      {
-        id: "zealous_crusade",
-        label: "Zealous Crusade",
-        aiPromptDirective:
-          "Write this character with intense, uncompromising conviction toward a specific dogma, deity, or ideal. Their speech is passionate, direct, and completely intolerant of compromise.",
-      },
-      {
-        id: "power_absolute",
-        label: "Power at All Costs",
-        aiPromptDirective:
-          "This character is entirely self-serving, ambitious, and ruthless. Their behavior can range from highly charismatic manipulation to terrifying authority, always prioritizing personal leverage.",
-      },
-    ],
-    "Cyberpunk / Corporate": [
-      {
-        id: "corporate_loyalist",
-        label: "Corporate Loyalist",
-        aiPromptDirective:
-          "Incorporate corporate buzzwords and clinical, risk-managed PR language. This character prioritizes systemic order, corporate policy, and lines of credit over human empathy.",
-      },
-      {
-        id: "street_pragmatist",
-        label: "Street Pragmatist",
-        aiPromptDirective:
-          "Use sharp, weary, and highly transactional street slang. This character has no illusions about saving the world; their focus is purely short-term survival and protecting their immediate crew.",
-      },
-      {
-        id: "ideological_radical",
-        label: "Ideological Radical",
-        aiPromptDirective:
-          "Write this character with raw, anti-establishment energy. They use rebellious, anti-corp rhetoric and are actively willing to burn down systems, entirely indifferent to collateral damage.",
-      },
-      {
-        id: "cold_professional",
-        label: "Cold Professional",
-        aiPromptDirective:
-          "Dialogue must be minimal, precise, and devoid of personal bias. The character views tasks purely as mechanical execution — nothing is personal, everything is just business.",
-      },
-      {
-        id: "burned_out_cynic",
-        label: "Burned-Out Cynic",
-        aiPromptDirective:
-          "Infuse the dialogue with deep nihilism, dry sarcasm, and exhaustion. The character has seen it all fall apart and operates on a baseline expectation of systemic failure.",
-      },
-      {
-        id: "predatory_opportunist",
-        label: "Predatory Opportunist",
-        aiPromptDirective:
-          "Make this character predatory, slippery, and untrustworthy. They use smooth talk to mask aggressive self-interest and are always scanning the room for an exploit or an escape route.",
-      },
-    ],
-    "Vampire / Gothic Noir": [
-      {
-        id: "strict_ascetic",
-        label: "Strict Ascetic",
-        aiPromptDirective:
-          "Write this character with a strained, hyper-controlled demeanor. They use precise, polite, and antiquated language as a psychological armor to contain an inner, volatile darkness.",
-      },
-      {
-        id: "haunted_sympathizer",
-        label: "Haunted Sympathizer",
-        aiPromptDirective:
-          "Dialogue should convey deep guilt, hesitation, and vulnerability. The character is deeply ashamed of their actions or status and is desperately looking for small ways to do good without getting caught.",
-      },
-      {
-        id: "cold_monster",
-        label: "Cold Monster",
-        aiPromptDirective:
-          "Write this character with a chilling, detached elegance. They lack basic human empathy, treating people like simple livestock or resources, completely free of guilt or malice.",
-      },
-      {
-        id: "obsessive_zealot",
-        label: "Obsessive Zealot",
-        aiPromptDirective:
-          "Infuse the character's descriptions with an eerie, manic focus. Their dialogue should repeatedly orient toward their singular, consuming fixation, overriding all other social cues.",
-      },
-      {
-        id: "decadent_hedonist",
-        label: "Decadent Hedonist",
-        aiPromptDirective:
-          "Dialogue is theatrical, cynical, and dripping with sensory indulgence. The character uses casual amusement and hedonism to aggressively deflect from their inner decay or emptiness.",
-      },
-      {
-        id: "pragmatic_survivor",
-        label: "Pragmatic Survivor",
-        aiPromptDirective:
-          "Write this NPC as highly defensive, paranoid, and survival-driven. Their speech is guarded, they avoid making definitive promises, and they prioritize exit strategies over loyalty.",
-      },
-    ],
-    "Sci-Fi / Space Opera": [
-      {
-        id: "system_loyalist",
-        label: "System Loyalist",
-        aiPromptDirective:
-          "This character speaks with the confidence of an institutionalist. They emphasize law, civilization, hierarchy, and data-driven stability, viewing rebels or independents as dangerous chaos.",
-      },
-      {
-        id: "frontier_independent",
-        label: "Frontier Independent",
-        aiPromptDirective:
-          "Use rugged, informal, and fiercely independent dialogue. The character values self-reliance, localized trust, and personal freedom above centralized planetary laws.",
-      },
-      {
-        id: "zealous_visionary",
-        label: "Zealous Visionary",
-        aiPromptDirective:
-          "Focus the character on the future, technology, or a grand cosmic mission. They view current human suffering or ethical concerns as insignificant speed bumps on the road to evolution.",
-      },
-      {
-        id: "principled_pacifist",
-        label: "Principled Pacifist",
-        aiPromptDirective:
-          "Dialogue must be calm, deeply humanistic, and actively seek compromise. The character maintains absolute ethical boundaries against violence, regardless of how harsh the setting is.",
-      },
-      {
-        id: "opportunistic_trader",
-        label: "Opportunistic Trader",
-        aiPromptDirective:
-          "Write this character with a highly commercial, speculative, and conversational tone. They treat every interaction as an open negotiation, constantly weighing cost-benefit ratios.",
-      },
-      {
-        id: "subversive_rebel",
-        label: "Subversive Rebel",
-        aiPromptDirective:
-          "The character operates with the secrecy and intensity of an active insurgent. Their language is revolutionary, defiant, and actively seeks to disrupt or dismantle structural authority.",
-      },
-    ],
-    "Modern Conspiracy": [
-      {
-        id: "institutionalist",
-        label: "Institutionalist",
-        aiPromptDirective:
-          "Use highly clinical, dry, and compartmentalized intelligence jargon. The character prioritizes official protocol, institutional security, and the chain of command above all else.",
-      },
-      {
-        id: "noble_transgressor",
-        label: "Noble Transgressor",
-        aiPromptDirective:
-          "Write this character as a quiet, hyper-focused operative who knowingly breaks laws to achieve a greater moral good. Their dialogue is guarded but deeply principled.",
-      },
-      {
-        id: "fanatical_believer",
-        label: "Fanatical Believer",
-        aiPromptDirective:
-          "The character's speech must carry the weight of dangerous, total certainty. They view all of human society as an illusion and treat people as mere assets or collateral to be spent for the Truth.",
-      },
-      {
-        id: "unprincipled_asset",
-        label: "Unprincipled Asset",
-        aiPromptDirective:
-          "Dialogue should be highly transactional, street-smart, and amoral. The character has zero ideological loyalty, viewing their skills and information purely as commodities to sell to the highest bidder.",
-      },
-      {
-        id: "haunted_insider",
-        label: "Haunted Insider",
-        aiPromptDirective:
-          "Infuse the character's tone with intense paranoia, panic, and deep ethical distress. They are physically and mentally exhausted from keeping terrible secrets and expect betrayal at any second.",
-      },
-      {
-        id: "machiavellian_player",
-        label: "Machiavellian Player",
-        aiPromptDirective:
-          "Write this character with a charming, highly collected, and disarming social facade. Beneath this exterior, their dialogue and actions are driven entirely by cold, calculated personal advancement.",
-      },
-    ],
-    "Post-Apocalyptic": [
-      {
-        id: "collectivist",
-        label: "Collectivist",
-        aiPromptDirective:
-          "This character speaks with a rugged, collective 'we' mentality. They prioritize the survival, defense, and material resources of their specific settlement over individual rights or outsiders.",
-      },
-      {
-        id: "tribal_xenophobe",
-        label: "Tribal Xenophobe",
-        aiPromptDirective:
-          "Write this character with intense suspicion, hostility, and localized, insular language. They view anyone outside their immediate clan as a lethal threat or an untrustworthy parasite.",
-      },
-      {
-        id: "pure_scavenger",
-        label: "Pure Scavenger",
-        aiPromptDirective:
-          "Dialogue is short, practical, and heavily focused on material scrap, ammo, and survival utilities. The character avoids any long-term commitments or alliances, relying entirely on themselves.",
-      },
-      {
-        id: "wasteland_zealot",
-        label: "Wasteland Zealot",
-        aiPromptDirective:
-          "Write this character with a strange, stylized fanaticism. They use bizarre vocabulary rooted in post-apocalyptic myths or cult beliefs, viewing the ruins of the world through a terrifying religious lens.",
-      },
-      {
-        id: "despotic_ruler",
-        label: "Despotic Ruler",
-        aiPromptDirective:
-          "The character's tone is authoritative, heavy, and threatening. They enforce their will through raw intimidation, justifying their cruelty as the only practical way to hold back total wasteland anarchy.",
-      },
-      {
-        id: "utopian_builder",
-        label: "Utopian Builder",
-        aiPromptDirective:
-          "Write this character with a resilient, hopeful, and idealistic tone. Despite the harsh wasteland environment, they emphasize laws, education, historical recovery, and long-term societal rebuilding.",
-      },
-    ],
-  } as Record<
-    string,
-    { id: string; label: string; aiPromptDirective: string }[]
-  >,
-};
-
-export const vampireConfig = {
-  archetypes: [
-    "Aristocratic Court",
-    "Occult Coven",
-    "Predatory Brood",
-    "Conspiring Syndicate",
-    "Rebel Anarchs",
-  ],
-  bloodlines: [
-    "Sanguine Nobles (Charismatic Mind-Benders)",
-    "Shadow Stalkers (Nightmare Weavers)",
-    "Blood Sorcerers (Occult Ritualists)",
-    "Bestial Ravagers (Feral Predator Shapeshifters)",
-    "Melancholic Artists (Aesthetes of Decay)",
-  ],
-  feedingHabits: [
-    "High-Society Salons (Elite & Consent-based)",
-    "Street Predation (Slums & Forgotten Alleys)",
-    "Blood Trafficking (Black Market & Clinics)",
-    "Occult Sacraments (Ritualistic & Sacrificial)",
-    "Wild Wilderness Hunts (Deep Forests & Ruins)",
-  ],
-  weaknesses: [
-    "Severe Sun Sensitivity (Burns instantly)",
-    "Consecrated Ground Aversion (Cannot cross thresholds)",
-    "Mirror & Reflection Absence (Exposes their nature)",
-    "Decaying Physical Form (Needs fresh blood to look human)",
-    "Silver & Wooden Vulnerability (Prevents regeneration)",
-  ],
-  scopes: [
-    "Single city underbelly",
-    "Hidden castle & border valley",
-    "Trade route shadow network",
-    "Metropolitan high society",
-    "Continental shadow court",
-  ],
-  alignments: [
-    "Strictly lawful, highly predatory",
-    "Pragmatic and power-driven",
-    "Feral, chaotic, and blood-fueled",
-    "Secretive and ritual-obsessed",
-    "Rebellious, seeking freedom from elders",
-  ],
-  goals: [
-    "Infiltrate the city council and blood-bind key mortal leaders.",
-    "Exhume the sarcophagus of their dormant Progenitor.",
-    "Monopolize the local blood bank distribution network.",
-    "Wipe out a rival werewolf pack or vampire hunters' cell.",
-    "Reconstruct a shattered ancient chronicle of blood magic.",
-  ],
-  conflicts: [
-    "The younger brood is planning a rebellion against the ancient elder.",
-    "A feeding gone wrong has drawn the attention of mortal authorities.",
-    "A rogue member has stolen a ledger detailing the clan's human farms.",
-    "The blood supply is tainted by a mystical pathogen.",
-    "An inquisitor has successfully tracked their primary haven.",
-  ],
-  hooks: [
-    "A mysterious patron hires the party to deliver a sealed urn, which is a vampire ashes decoy.",
-    "Locals ask the party to investigate a series of bloodless bodies found in the canal.",
-    "A member of the clan offers to sell a list of high-profile vampire thralls in the city council.",
-    "The clan captures the party and offers freedom in exchange for retrieving a relic from a sunlit temple.",
-    "A dying vampire begs the party to protect their mortal family from their own sire.",
-  ],
-};
-
-export const questConfig = {
-  genres: [
-    "Classic Fantasy",
-    "Dark Fantasy",
-    "Political Intrigue",
-    "Horror",
-    "Comedy",
-  ],
-  tones: ["Heroic", "Gritty", "Mysterious", "Comedic", "Tragic"],
-  scopes: [
-    "Local (village / district)",
-    "Regional (kingdom / region)",
-    "World-threatening",
-  ],
-  locationTypes: [
-    "Ancient Dungeon",
-    "Urban City",
-    "Wilderness",
-    "Cursed Ruin",
-    "Coastal / Maritime",
-    "Planar Realm",
-  ],
-  threats: [
-    "Monstrous Creature",
-    "Corrupt Villain",
-    "Rival Faction",
-    "Ancient Curse",
-    "Natural Disaster",
-    "Betrayal from Within",
-  ],
-  hooks: [
-    "A local official offers a reward to find a missing heir before a rival claims the title.",
-    "Strange lights above the old tower have kept the village awake for three nights.",
-    "A caravan was found empty on the road — no blood, no tracks, just silence.",
-    "An imprisoned criminal offers the location of a cache in exchange for a pardon.",
-    "A temple guardian collapses mid-ceremony, whispering a single forbidden name.",
-    "A child wanders into town carrying an item that should not exist.",
-  ],
-  complications: [
-    "The client is hiding their true motive — the real target is someone the party knows.",
-    "The threat has already moved; the location the party was sent to is a decoy.",
-    "A second party has been hired for the same job and has a head start.",
-    "The threat is connected to a powerful patron who expects the party to look away.",
-    "Completing the job requires breaking a law the party has respected until now.",
-    "The reward will not be paid unless the job is done silently and officially denied.",
-  ],
-  twists: [
-    "Villain is protecting something valuable",
-    "Client's ally is the real enemy",
-    "Location holds a campaign-changing secret",
-    "Threat must be bargained with, not killed",
-    "Party's own past caused this situation",
-    "Two factions both claim the prize",
-  ],
-  rewards: [
-    "Coin plus a local power's favor",
-    "Deed to a useful property",
-    "Access to a restricted archive",
-    "A magic item from the site",
-    "Valuable information from the client",
-    "Respect from a previously hostile faction",
-  ],
-};
-
-export const socialHubConfig = {
-  genres: [
-    "Fantasy",
-    "Dark Fantasy",
-    "Pirate",
-    "Cyberpunk",
-    "Sci-Fi",
-    "Modern",
-    "Horror",
-    "Post-Apocalyptic",
-    "Western",
-  ],
-  venueTypesByGenre: {
-    Fantasy: [
-      "Tavern / Inn",
-      "Mead Hall",
-      "Roadside Alehouse",
-      "Adventurer Lodge",
-      "Guildhall",
-    ],
-    "Dark Fantasy": [
-      "Cursed Alehouse",
-      "Witch's Den",
-      "Underground Fighting Pit",
-      "Plague Hospice",
-      "Smuggler's Hollow",
-    ],
-    Pirate: [
-      "Dockside Tavern",
-      "Rum House",
-      "Sailor's Inn",
-      "Gambling Den",
-      "Freeport Alehouse",
-    ],
-    Cyberpunk: [
-      "Noodle Bar",
-      "Dive Bar",
-      "Nightclub",
-      "Hacker Café",
-      "Braindance Lounge",
-    ],
-    "Sci-Fi": [
-      "Spaceport Cantina",
-      "Station Bar",
-      "Mess Hall",
-      "Orbital Lounge",
-      "Asteroid Miner Pub",
-    ],
-    Modern: ["Pub", "Café / Diner", "Hotel Bar", "Nightclub", "Truck Stop"],
-    Horror: [
-      "Goth Club",
-      "Occult Café",
-      "Speakeasy",
-      "Blood Bar",
-      "Private Lounge",
-    ],
-    "Post-Apocalyptic": [
-      "Trade Shack",
-      "Bunker Canteen",
-      "Water Bar",
-      "Settlement Mess",
-      "Caravanserai",
-    ],
-    Western: [
-      "Saloon",
-      "Boarding House",
-      "Trading Post",
-      "Roadhouse",
-      "Gambling Parlour",
-    ],
-  } as Record<string, string[]>,
-  atmospheres: [
-    "Rowdy and welcoming",
-    "Tense and suspicious",
-    "Quiet and melancholic",
-    "Festive and chaotic",
-    "Cold and professional",
-    "Warm but secretive",
-  ],
-  wealthLevels: [
-    "Destitute (dirt floors, watered-down drinks)",
-    "Poor (cheap but honest)",
-    "Modest (reliable, no frills)",
-    "Comfortable (decent food and beds)",
-    "Prosperous (good drink, private rooms)",
-    "Wealthy (exclusive clientele)",
-  ],
-  clientelesByGenre: {
-    Fantasy: [
-      "Adventurers and wanderers",
-      "Merchants and traders",
-      "Soldiers and mercenaries",
-      "Pilgrims and clergy",
-      "Criminals and fence-seekers",
-      "Mixed locals",
-    ],
-    "Dark Fantasy": [
-      "Mercenaries and cultists",
-      "Outlaws and desperate souls",
-      "Corrupt clergy",
-      "Cursed travellers",
-      "Black-market traders",
-    ],
-    Pirate: [
-      "Pirates and privateers",
-      "Sailors and dockworkers",
-      "Smugglers and fences",
-      "Bounty hunters",
-      "Stranded merchants",
-    ],
-    Cyberpunk: [
-      "Hackers and netrunners",
-      "Off-duty security",
-      "Smugglers and fixers",
-      "Street gang members",
-      "Corporate burnouts",
-    ],
-    "Sci-Fi": [
-      "Spacers and pilots",
-      "Colonial marines",
-      "Free traders",
-      "Scientists and researchers",
-      "Station workers",
-    ],
-    Modern: [
-      "Office workers",
-      "Local regulars",
-      "Tourists and travellers",
-      "Journalists and students",
-      "Off-duty police",
-    ],
-    Horror: [
-      "Occultists and hunters",
-      "Lost souls and drifters",
-      "Curious investigators",
-      "Predators in disguise",
-      "Frightened locals",
-    ],
-    "Post-Apocalyptic": [
-      "Scavengers and traders",
-      "Wasteland survivors",
-      "Cult members",
-      "Mercenaries",
-      "Settlers and refugees",
-    ],
-    Western: [
-      "Cowboys and drifters",
-      "Miners and prospectors",
-      "Outlaws and bounty hunters",
-      "Townsfolk",
-      "Railroad workers",
-    ],
-  } as Record<string, string[]>,
-  troubles: [
-    "The owner owes a dangerous debt that is coming due",
-    "A recent violent incident was quietly buried — the responsible party may still be inside",
-    "A protected criminal is hiding among the staff",
-    "A back room connects to something the owner refuses to discuss",
-    "A faction is using the venue as a dead-drop without the owner's knowledge",
-    "The place is being squeezed out by a rival backed by local power",
-  ],
-  settlementTypes: [
-    "Capital city",
-    "Market town",
-    "Frontier outpost",
-    "Coastal port",
-    "Remote village",
-    "Crossroads hamlet",
-  ],
-};
-
-export const nationConfig = {
-  genres: [
-    "Fantasy",
-    "Dark Fantasy",
-    "Pirate",
-    "Cyberpunk",
-    "Sci-Fi",
-    "Modern",
-    "Horror",
-    "Post-Apocalyptic",
-    "Western",
-  ],
-  polityTypesByGenre: {
-    Fantasy: [
-      "Kingdom",
-      "Empire",
-      "City-State",
-      "Duchy",
-      "Theocracy",
-      "Tribal Confederation",
-      "Republic",
-      "Protectorate",
-    ],
-    "Dark Fantasy": [
-      "Cursed Realm",
-      "Necromancer's Domain",
-      "Warlord Territory",
-      "Fallen Empire",
-      "Plague State",
-    ],
-    Pirate: [
-      "Free Port",
-      "Pirate Confederation",
-      "Merchant Republic",
-      "Privateer Kingdom",
-      "Smuggler's Haven",
-    ],
-    Cyberpunk: [
-      "Megacorp-State",
-      "Gang Territory",
-      "Free Zone",
-      "Corporate Enclave",
-      "Tech Oligarchy",
-    ],
-    "Sci-Fi": [
-      "Interstellar Federation",
-      "Colony Authority",
-      "Trade Consortium",
-      "Military Junta",
-      "Hive Mind Territory",
-    ],
-    Modern: [
-      "Nation-State",
-      "Autonomous Region",
-      "Rogue State",
-      "International Bloc",
-      "Failed State",
-    ],
-    Horror: [
-      "Cursed Domain",
-      "Cult Territory",
-      "Isolated Community",
-      "Shadow Government",
-      "Quarantine Zone",
-    ],
-    "Post-Apocalyptic": [
-      "Warlord Territory",
-      "Survivor Settlement",
-      "Cult Domain",
-      "Scavenger Confederation",
-      "Bunker State",
-    ],
-    Western: [
-      "Territory",
-      "Frontier Confederation",
-      "Railroad Authority",
-      "Outlaw Republic",
-      "Native Nation",
-    ],
-  } as Record<string, string[]>,
-  governmentStyles: [
-    "Hereditary dynasty",
-    "Elected council",
-    "Military dictatorship",
-    "Theocratic rule",
-    "Oligarchy",
-    "Constitutional government",
-    "Tribal / clan elders",
-    "Corporate board",
-    "Revolutionary committee",
-  ],
-  scales: [
-    "City-state (single settlement)",
-    "Small region (handful of towns)",
-    "Mid-sized nation",
-    "Large empire (many provinces)",
-    "Continent-spanning dominion",
-  ],
-  conflictLevels: [
-    "At peace",
-    "Simmering tensions",
-    "Open conflict",
-    "Total war",
-    "Post-war recovery",
-  ],
-  troubles: [
-    "The ruling class is fracturing over a succession crisis",
-    "A powerful faction is quietly positioning to seize control",
-    "An economic collapse is being hidden from the public",
-    "A border dispute is being deliberately escalated by a third party",
-    "A secret the ruler is keeping could delegitimise the entire state",
-    "An internal resistance movement is closer to open revolt than anyone admits",
-  ],
-};
-
-export const kingdomConfig = {
-  polityTypes: nationConfig.polityTypesByGenre["Fantasy"],
-  governmentStyles: nationConfig.governmentStyles.slice(0, 7),
-  geographies: [
-    "Temperate highlands",
-    "Coastal plains",
-    "Dense forest",
-    "Arid desert",
-    "Arctic tundra",
-    "Island archipelago",
-    "River delta",
-    "Mountain range",
-  ],
-  scales: nationConfig.scales,
-  conflictLevels: nationConfig.conflictLevels,
-  magicLevels: [
-    "No magic",
-    "Rare and feared",
-    "Common but regulated",
-    "Widely practiced",
-    "Magic-dependent society",
-  ],
-  troubles: nationConfig.troubles,
-};
-
-export interface GeneratorOutput {
-  type:
-    | "character"
-    | "creature"
-    | "location"
-    | "item"
-    | "event"
-    | "faction"
-    | "note";
-  title: string;
-  summary?: string;
-  content: string;
-  lore: string;
-  labels: string[];
-  status: "active" | "draft";
-}
+import { generateName as _generateName } from "./generators/base";
+import { generateNPC } from "./generators/npc";
+import { generateFaction, generateVampireClan } from "./generators/faction";
+import { generateSettlement } from "./generators/settlement";
+import { generateMagicItem } from "./generators/magic-item";
+import { generateQuestHook } from "./generators/quest";
+import { generateNames } from "./generators/names";
+import { generateSocialHub, generateTavern } from "./generators/social-hub";
+import { generateKingdom, generateNation } from "./generators/kingdom-nation";
+import type { GeneratorOutput } from "./generators/base";
 
 export class DefaultGeneratorEngine {
   constructor(private clientManager = aiClientManager) {}
 
-  /**
-   * Generates a random name based on prefixes and suffixes.
-   */
   generateName(): string {
-    const prefix =
-      nameTable.prefixes[Math.floor(Math.random() * nameTable.prefixes.length)];
-    const suffix =
-      nameTable.suffixes[Math.floor(Math.random() * nameTable.suffixes.length)];
-    const useDescriptor = Math.random() > 0.6;
-    let name = `${prefix}${suffix}`;
-    if (useDescriptor) {
-      const descriptor =
-        nameTable.descriptors[
-          Math.floor(Math.random() * nameTable.descriptors.length)
-        ];
-      name = `${name} ${descriptor}`;
-    }
-    return name;
+    return _generateName();
   }
 
-  /**
-   * Generates an NPC draft.
-   */
   async generateNPC(
-    options: {
-      race?: string;
-      ancestry?: string;
-      role?: string;
-      alignment?: string;
-      campaignContext?: string;
-      theme?: string;
-      useAI?: boolean;
-    } = {},
+    options: Parameters<typeof generateNPC>[1] = {},
   ): Promise<GeneratorOutput> {
-    const theme = options.theme;
-    const race =
-      options.ancestry ||
-      options.race ||
-      npcConfig.races[Math.floor(Math.random() * npcConfig.races.length)];
-    const role =
-      options.role ||
-      npcConfig.roles[Math.floor(Math.random() * npcConfig.roles.length)];
-    const alignment =
-      options.alignment ||
-      npcConfig.alignments[
-        Math.floor(Math.random() * npcConfig.alignments.length)
-      ];
-    const campaignContext = options.campaignContext?.trim();
-    const name = this.generateName();
-
-    const npcNamingStyles = [
-      "Give the NPC a name that sounds distinctly local to their culture — not generic fantasy.",
-      "Use a name with unusual phonetic texture. Avoid Kael, Zara, Theron, Vane, Kane, Drake, Stone, Grey, Ash, Cole, and similar overused patterns.",
-      "Give the NPC a short epithet or title that hints at their reputation — invent an original one, do not reuse common examples.",
-      "Use a name that suggests a specific cultural or ethnic origin consistent with their ancestry.",
-      "Choose a name that is easy to say aloud at a gaming table — short, distinct, memorable, and not a common English surname.",
-    ];
-    const chosenNamingStyle = this.pickFrom(npcNamingStyles);
-    const varianceSeed = Math.floor(Math.random() * 99991) + 10;
-
-    if (options.useAI !== false) {
-      try {
-        const npcThemeVoice: Record<string, string> = {
-          "Classic Fantasy":
-            "medieval fantasy — guilds, nobles, arcane orders, political intrigue in a world of swords and sorcery",
-          "Cyberpunk / Corporate":
-            "near-future cyberpunk — megacorporations, street gangs, hackers, corporate espionage, neon-lit dystopia",
-          "Vampire / Gothic Noir":
-            "gothic horror — vampire covens, inquisitions, decadent aristocracy, forbidden rites, candlelit conspiracies",
-          "Sci-Fi / Space Opera":
-            "science fiction space opera — stellar federations, alien factions, interstellar trade, colony politics, advanced technology",
-          "Modern Conspiracy":
-            "modern-day thriller — intelligence agencies, secret societies, corporate conspiracies, hidden influence networks",
-          "Post-Apocalyptic":
-            "post-apocalyptic survival — scavenger tribes, wasteland cults, resource wars, collapsed civilisation, desperate factions",
-        };
-        const voice = theme
-          ? (npcThemeVoice[theme] ?? "tabletop RPG")
-          : "tabletop RPG";
-
-        const systemInstruction = `You are an expert RPG campaign writer specialising in ${voice}. You generate detailed, original NPC drafts for that setting in JSON format.
-
-OUTPUT FORMAT — return ONLY a valid JSON object, no markdown fences:
-{
-  "title": "NPC name (follow the naming directive in the user message)",
-  "summary": "One sentence: who this NPC is and what makes them interesting (e.g. 'A disgraced noble archivist who sells secrets to fund a private obsession.').",
-  "content": "Markdown. Use exactly these four section headers in order: '### Who they are', '### What they want', '### Why they are useful', '### How to use them at the table'. Each section: 2-4 tight sentences. Include campaign context if provided.",
-  "lore": "Markdown. Use EXACTLY this structure with ### headers and '- **Label**: Value' list items:\\n### At a Glance\\n- **Ancestry**: race and background\\n- **Role**: what they do\\n- **Moral Stance**: behavioral anchor\\n- **Secret**: hidden truth that would change everything\\n- **Immediate Hook**: one-sentence GM hook\\n### Personality\\n- two distinct personality traits as bullet points\\n### Faction Connection\\none sentence on their organisational ties or lack thereof",
-  "labels": ["2-4 lowercase tags describing their role and traits, plus 'rpg-character', 'npc-generator', 'imported-draft'"]
-}
-
-QUALITY RULES:
-- Every NPC must feel like a completely different person — avoid repeating names, archetypes, or backstory structures.
-- Avoid overused NPC name patterns. Do NOT use: Kael, Theron, Zara, Aldric, Vane, Kane, Drake, Stone, Grey, Ash, Cole, Maren, Cross, Vale, or common English monosyllable surnames.
-- The secret should be genuinely surprising and table-usable, not a generic "dark past."
-- Place names must be specific and invented — no "the old district", "the lower city", "Oakhaven", "Millbrook", "Riverdale", or similar generic compound settlement names.
-- Before finalising, silently check for: name not on the forbidden list; secret is genuinely surprising and not contradicted by the stated role or faction connection; all four content sections are internally consistent (what they want should explain why they are useful; their secret should reframe who they are). Rewrite any section where a contradiction exists.`;
-
-        const moralityAnchor = theme
-          ? npcThemeConfig.moralities[theme]?.find((m) => m.id === alignment)
-          : undefined;
-        const behavioralDirective =
-          moralityAnchor?.aiPromptDirective ?? alignment;
-        const moralityLabel = moralityAnchor?.label ?? alignment;
-
-        const userMessage = `Generate an NPC. Variation seed: ${varianceSeed}.
-${theme ? `- Genre/Theme: ${theme}` : ""}
-- Ancestry/Race: ${race}
-- Role: ${role}
-- Moral Stance: ${moralityLabel}
-- Behavioral Directive: ${behavioralDirective}${campaignContext ? `\n- Campaign Context: ${campaignContext}` : ""}
-- Naming Directive: ${chosenNamingStyle}`;
-
-        const model = await this.clientManager.getModel(
-          "",
-          "gemini-3.1-flash-lite",
-          systemInstruction,
-        );
-        const response = await model.generateContent(userMessage);
-        const text = response.response.text().trim();
-        const cleanText = text
-          .replace(/^```json\s*/i, "")
-          .replace(/```$/, "")
-          .trim();
-        const data = JSON.parse(cleanText);
-
-        return {
-          type: "character",
-          title: data.title || name,
-          summary:
-            data.summary ||
-            `A ${moralityLabel.toLowerCase()} ${race.toLowerCase()} ${role.toLowerCase()} with something to hide.`,
-          content: data.content || "",
-          lore: data.lore || "",
-          labels: Array.isArray(data.labels)
-            ? data.labels
-            : ["rpg-character", "npc-generator", "imported-draft"],
-          status: "active",
-        };
-      } catch (err) {
-        console.warn(
-          "AI generation failed, falling back to local tables:",
-          err,
-        );
-      }
-    }
-
-    // Local Deterministic/Random Fallback
-    const traits = this.getRandomItems(npcConfig.traits, 2);
-    const secret =
-      npcConfig.secrets[Math.floor(Math.random() * npcConfig.secrets.length)];
-    const motive =
-      npcConfig.motives[Math.floor(Math.random() * npcConfig.motives.length)];
-    const faction =
-      npcConfig.factions[Math.floor(Math.random() * npcConfig.factions.length)];
-    const plotHook =
-      npcConfig.plotHooks[
-        Math.floor(Math.random() * npcConfig.plotHooks.length)
-      ];
-
-    const fallbackMoralityLabel = theme
-      ? (npcThemeConfig.moralities[theme]?.find((m) => m.id === alignment)
-          ?.label ?? alignment)
-      : alignment;
-    const summary = `A ${fallbackMoralityLabel.toLowerCase()} ${race.toLowerCase()} ${role.toLowerCase()} with something to hide.`;
-
-    const content = `### Who they are
-${name} is a ${race} ${role} whose public reputation is useful, incomplete, and just suspicious enough to matter. Locals know them as someone who gets results, even when the work requires favors, secrets, or a carefully timed lie.${campaignContext ? ` In ${campaignContext}, they are already entangled in the edges of the main conflict.` : ""}
-
-### What they want
-${motive} Everything they do, however helpful it appears on the surface, is filtered through this underlying drive.
-
-### Why they are useful
-${faction} They know routes, names, prices, and debts that the party cannot easily learn any other way.
-
-### How to use them at the table
-Introduce ${name} when the party needs a social lead, a compromised witness, or a morally complicated ally. They should be helpful immediately — but never free of consequences.`;
-
-    const lore = `### At a Glance
-- **Ancestry**: ${race}
-- **Role**: ${role}
-- **Moral Stance**: ${fallbackMoralityLabel}
-- **Secret**: ${secret}
-- **Immediate Hook**: ${plotHook}
-
-### Personality
-- ${traits[0]}
-- ${traits[1]}
-
-### Faction Connection
-${faction}`;
-
-    return {
-      type: "character",
-      title: name,
-      summary,
-      content,
-      lore,
-      labels: ["rpg-character", "npc-generator", "imported-draft"],
-      status: "active",
-    };
+    return generateNPC(this.clientManager, options);
   }
 
-  private pickFrom<T>(arr: T[]): T {
-    return arr[Math.floor(Math.random() * arr.length)];
-  }
-
-  private factionBase(type: string): string {
-    const map: Record<string, string[]> = {
-      "Merchant Guild": [
-        "A bonded counting house whose ledgers are sealed by city charter",
-        "A licensed exchange hall at the centre of the trade district",
-        "A warehouse compound that no sheriff may enter without a writ",
-      ],
-      "Secret Society": [
-        "A private dining club whose membership list is never committed to paper",
-        "A decommissioned observatory reached through a hidden press in the library stacks",
-        "Rotating safe houses connected by messenger-drop protocols",
-      ],
-      "Mercenary Company": [
-        "A fortified barracks compound outside the city walls",
-        "A charted garrison holding neutral ground between two rival lords",
-        "A licensed inn that doubles as a staging ground for contract work",
-      ],
-      "Temple Order": [
-        "A sanctified compound built above the sealed catacombs",
-        "A pilgrimage waystation that doubles as an intelligence hub",
-        "A charitable hospice whose basement holds restricted archives",
-      ],
-      "Criminal Syndicate": [
-        "A legitimate bathhouse with soundproofed rooms below street level",
-        "A moneylender's office whose public ledgers contain a second set of books",
-        "A district of connected properties linked by sealed passages",
-      ],
-      "Rebel Cell": [
-        "A print-house running two sets of accounts",
-        "A disused chapel in a contested neighbourhood where records are rarely checked",
-        "A network of sympathiser homes linked by a rotating code phrase",
-      ],
-      "Arcane Circle": [
-        "A registered scholar's hall with warded inner chambers",
-        "A cartographer's guild whose maps contain hidden notation systems",
-        "A canal barge anchored in a dock district where manifests go uninspected",
-      ],
-      // Cyberpunk types
-      "Megacorporation Megagroup": [
-        "A sealed corporate tower whose lower floors are open to the public and upper floors are not on any map",
-        "A campus of linked facilities connected by private transit lines that bypass city checkpoints",
-        "A data-centre compound in a legally ambiguous special economic zone",
-      ],
-      "Corporate Syndicate": [
-        "A registered LLC with rotating directors and no fixed address",
-        "A licensed private security firm that maintains offices in three jurisdictions simultaneously",
-        "A shell company whose registered seat is a post-box in a compliant offshore district",
-      ],
-      "Hacker Collective": [
-        "A distributed mesh of rented server nodes and anonymous relay points",
-        "A legitimate ISP whose routing infrastructure doubles as a covert comms layer",
-        "Rotating physical dead-drops in public infrastructure — lockers, charging stations, transit hubs",
-      ],
-      "Street Gang Alliance": [
-        "A block of contested commercial units enforced by informal tax agreements",
-        "A series of interconnected basement spaces beneath a market district",
-        "A community centre operating with city permits while the basement handles other business",
-      ],
-      // Gothic types
-      "Vampire Coven": [
-        "A sealed private estate whose deed has not changed hands in three centuries",
-        "A licensed sanatorium whose patient records are never released to outside authorities",
-        "A labyrinthine wine cellar beneath a respectable merchant's townhouse",
-      ],
-      "Inquisition Watch": [
-        "A fortified chapter-house adjacent to the civil courthouse",
-        "A mobile tribunal that establishes temporary jurisdiction wherever the investigation leads",
-        "A warded archive annexed to the city's oldest cathedral",
-      ],
-      // Sci-Fi types
-      "Stellar Federation Alliance": [
-        "A neutral space station positioned at a strategically contested transit point",
-        "A diplomatic compound on a contested colony world with extraterritorial status",
-        "A fleet of registered humanitarian vessels that doubles as a mobile command structure",
-      ],
-      // Modern types
-      "Intelligence Agency": [
-        "A nondescript government office building whose basement floors are not on the building plan",
-        "A chain of legitimate consulting firms that share encrypted back-office infrastructure",
-        "An embassy annex operating under diplomatic immunity",
-      ],
-      // Post-apocalyptic types
-      "Scavenger Tribe": [
-        "A fortified salvage yard at the edge of a collapsed industrial zone",
-        "A mobile convoy that claims no fixed territory but controls key supply corridors",
-        "A series of hidden caches spread across a hundred kilometres of dead highway",
-      ],
-      "Wasteland Cult": [
-        "A sealed compound built inside a pre-collapse water treatment facility",
-        "A fortified hilltop site with sightlines across three days of travel in every direction",
-        "A network of underground bunkers connected by service tunnels from before the collapse",
-      ],
-    };
-    return this.pickFrom(
-      map[type] ?? [
-        "A neutral facility whose access is controlled and whose records are not shared",
-        "A licensed premises that provides cover for activities conducted elsewhere",
-        "A distributed network of locations with no single point of failure",
-      ],
-    );
-  }
-
-  private factionResource(type: string): string {
-    const map: Record<string, string[]> = {
-      "Merchant Guild": [
-        "Exclusive trade licences, bonded debts, and letters of introduction that open every city gate",
-        "Commodity price information days before it reaches the open market",
-        "Certified seals of provenance that determine what goods may legally change hands",
-      ],
-      "Secret Society": [
-        "Compromising knowledge distributed in sealed fragments held by separate members",
-        "A curated register of favours owed by officials, merchants, and clergy",
-        "Access to a network of false identities and safe-passage routes",
-      ],
-      "Mercenary Company": [
-        "Contractual access to trained soldiers who ask no political questions",
-        "Neutral enforcement services hired by every side of every dispute",
-        "An archive of battlefield contracts that constitute decades of political leverage",
-      ],
-      "Temple Order": [
-        "Exclusive rights over burial rites, confessions, and civic oaths",
-        "A pharmaceutical supply chain running through the charitable district",
-        "Institutional immunity protecting their premises from search or seizure",
-      ],
-      "Criminal Syndicate": [
-        "Control over the city's informal credit markets and enforcement ecosystem",
-        "Detailed knowledge of every patrol route, informant, and magistrate's price",
-        "A distribution network for restricted goods running through legitimate storefronts",
-      ],
-      "Rebel Cell": [
-        "A verified printing and distribution network for prohibited materials",
-        "Contacts embedded in the guard, the census office, and the merchant registry",
-        "Secure courier routes that move people, messages, and contraband past checkpoints",
-      ],
-      "Arcane Circle": [
-        "Proprietary ritual techniques licensed to no outside practitioner",
-        "A sealed archive of magical precedents that defines what is legally permitted",
-        "Controlled access to rare components that no other supplier will touch",
-      ],
-      // Cyberpunk types
-      "Megacorporation Megagroup": [
-        "Patent portfolios, regulatory capture, and the ability to rewrite local law through lobbying",
-        "A private security force larger than the city police and legally permitted to operate with fewer constraints",
-        "Exclusive contracts with critical infrastructure — power, water, data, transit",
-      ],
-      "Corporate Syndicate": [
-        "Shell-company ownership of key residential and commercial properties across the district",
-        "Leveraged debt held against every small business in the target sector",
-        "Proprietary logistics infrastructure that competitors cannot access without their permission",
-      ],
-      "Hacker Collective": [
-        "Zero-day exploits, surveillance backdoors, and access to every networked system in the city",
-        "A distributed archive of intercepted communications from every major institution",
-        "The ability to make anyone's digital identity disappear — or reappear differently",
-      ],
-      "Street Gang Alliance": [
-        "Control of informal economies: protection, distribution, and dispute resolution in three districts",
-        "Detailed knowledge of every surveillance blind spot, patrol schedule, and officer price",
-        "Loyalty networks that extend into city maintenance, transit, and low-level civil service",
-      ],
-      // Gothic types
-      "Vampire Coven": [
-        "Centuries of accumulated wealth, property, and blackmail material on every notable family",
-        "The ability to alter memory, compel testimony, and move unseen through any social tier",
-        "A network of thralls embedded in the city's legal, medical, and religious institutions",
-      ],
-      "Inquisition Watch": [
-        "Legal authority to detain, interrogate, and seize assets without civil court oversight",
-        "An archive of confessions, heresies, and crimes dating back three generations",
-        "Jurisdiction that supersedes local law in matters defined — broadly — as spiritual threat",
-      ],
-      // Sci-Fi types
-      "Stellar Federation Alliance": [
-        "Trade route licensing, customs authority, and the right to impose blockades under federation charter",
-        "A shared military asset pool that member states cannot individually match",
-        "Diplomatic recognition that determines which colonies and stations are treated as sovereign",
-      ],
-      // Modern types
-      "Intelligence Agency": [
-        "Surveillance infrastructure covering communications, financial transactions, and physical movement",
-        "Classified leverage on every significant political, corporate, and criminal actor in the region",
-        "The legal authority to classify, redact, and deny — which is effectively the power to erase events",
-      ],
-      // Post-apocalyptic types
-      "Scavenger Tribe": [
-        "Access to pre-collapse technology caches and the knowledge to operate what others cannot",
-        "Control of the only reliable route through a stretch of dead territory",
-        "A repair and fabrication capability that no other group in the region can match",
-      ],
-      "Wasteland Cult": [
-        "Clean water, food stockpiles, and medical supplies — distributed exclusively to the faithful",
-        "A coherent ideology that provides meaning in a world without institutions",
-        "Armed enforcers who believe completely in what they are protecting",
-      ],
-    };
-    return this.pickFrom(
-      map[type] ?? [
-        "Specialised knowledge or access that no other group in the region controls",
-        "A network of obligations, debts, and dependencies too entangled to cut cleanly",
-        "Control of a single critical resource that everyone else needs to function",
-      ],
-    );
-  }
-
-  /**
-   * Generates a faction draft.
-   */
   async generateFaction(
-    options: {
-      type?: string;
-      scope?: string;
-      alignment?: string;
-      campaignContext?: string;
-      theme?: string;
-      useAI?: boolean;
-    } = {},
+    options: Parameters<typeof generateFaction>[1] = {},
   ): Promise<GeneratorOutput> {
-    const theme = options.theme || factionConfig.themes[0];
-    const factionType =
-      options.type ||
-      factionConfig.types[
-        Math.floor(Math.random() * factionConfig.types.length)
-      ];
-    const scope =
-      options.scope ||
-      factionConfig.scopes[
-        Math.floor(Math.random() * factionConfig.scopes.length)
-      ];
-    const alignment =
-      options.alignment ||
-      factionConfig.alignments[
-        Math.floor(Math.random() * factionConfig.alignments.length)
-      ];
-    const campaignContext = options.campaignContext?.trim();
-    const name = `${this.generateName()} Compact`;
-
-    const namingStyles = [
-      "Name this faction after a material, substance, or natural phenomenon twisted to their purpose.",
-      "Name this faction after an abstract concept, virtue, or doctrine — not a person or place.",
-      "Use a short stark one-word name or a tight two-word compound (e.g. 'The Writ', 'Iron Accord').",
-      "Base the faction name on a specific local landmark, street, district, or geographic feature.",
-      "Name the faction after their founding secret, hidden method, or signature act.",
-      "Use a name that sounds like a legitimate civic institution but carries a sinister undertone.",
-      "Name the faction after a historical event, failed uprising, or forgotten figure from the setting.",
-      "Give the faction a name derived from an unusual profession, trade, or craft.",
-      "Use an archaic or invented word that evokes the faction's cultural roots.",
-      "Name the faction after a symbol, emblem, or recurring motif associated with their work.",
-    ];
-    const npcNamingStyles = [
-      "Give each NPC a name that sounds distinctly local — not generic fantasy.",
-      "Each NPC name should have an unusual phonetic texture. Avoid Kael, Zara, Theron, Vane, Kane, Drake, Stone, Grey, Ash, Cole, and similar overused patterns.",
-      "Give each NPC a short street name or title that hints at their role — invent an original one, do not reuse common examples.",
-      "Use names that suggest a specific cultural or ethnic origin consistent with the setting.",
-      "Each NPC should have a name that is easy to say aloud at a gaming table.",
-    ];
-    const chosenNamingStyle =
-      namingStyles[Math.floor(Math.random() * namingStyles.length)];
-    const chosenNpcStyle =
-      npcNamingStyles[Math.floor(Math.random() * npcNamingStyles.length)];
-    const varianceSeed = Math.floor(Math.random() * 99991) + 10;
-
-    if (options.useAI !== false) {
-      try {
-        const themeVoice: Record<string, string> = {
-          "Classic Fantasy":
-            "medieval fantasy — guilds, nobles, arcane orders, political intrigue in a world of swords and sorcery",
-          "Cyberpunk / Corporate":
-            "near-future cyberpunk — megacorporations, street gangs, hackers, corporate espionage, neon-lit dystopia",
-          "Vampire / Gothic Noir":
-            "gothic horror — vampire covens, inquisitions, decadent aristocracy, forbidden rites, candlelit conspiracies",
-          "Sci-Fi / Space Opera":
-            "science fiction space opera — stellar federations, alien factions, interstellar trade, colony politics, advanced technology",
-          "Modern Conspiracy":
-            "modern-day thriller — intelligence agencies, secret societies, corporate conspiracies, hidden influence networks",
-          "Post-Apocalyptic":
-            "post-apocalyptic survival — scavenger tribes, wasteland cults, resource wars, collapsed civilisation, desperate factions",
-        };
-        const voice = themeVoice[theme] ?? "tabletop RPG";
-
-        const systemInstruction = `You are an expert RPG campaign writer specialising in ${voice}. You generate detailed, original faction drafts for that setting in JSON format.
-
-OUTPUT FORMAT — return ONLY a valid JSON object, no markdown fences:
-{
-  "title": "Faction name (follow the naming directive in the user message)",
-  "summary": "One sentence: what this faction is and what makes them interesting (e.g. 'A sanitation cult-technocracy that controls clean water in a poisoned city.').",
-  "content": "Markdown. Use exactly these four section headers in order: '### What they control', '### What they want', '### Why they are dangerous', '### How to use them at the table'. Each section: 2-4 tight sentences. Include campaign context if provided.",
-  "lore": "Markdown. Use EXACTLY this structure with ### headers and '- **Label**: Value' list items:\\n### At the Table\\n- **Base**: specific named location\\n- **Resource**: what they control that others need\\n- **Symbol**: identifying mark or emblem\\n- **Secret**: hidden truth that would destroy them\\n- **Immediate Hook**: one-sentence GM hook\\n### Notable NPCs\\n- **Name**: one-line description (2-3 NPCs)\\n### Internal Conflict\\none paragraph\\n### Rival Faction\\n- **Name**: one-line rivalry",
-  "labels": ["2-5 lowercase tags for the faction's theme and activities, plus 'rpg-faction', 'faction-generator', 'imported-draft'"]
-}
-
-QUALITY RULES:
-- Every generation must feel like a completely different faction — avoid repeating names, concepts, or structures from prior outputs.
-- Avoid generic RPG naming clichés (no 'Gilded Ledger', 'Iron Brotherhood', 'Shadow Hand', etc.).
-- NPC names must feel culturally specific and phonetically varied — do NOT use Kael, Zara, Theron, Vane, Kane, Drake, Stone, Grey, Ash, Cole, Maren, Cross, Vale, or common English monosyllable surnames.
-- Place names (bases, districts, landmarks) must be specific and invented — never use 'the old district', 'the lower city', 'Oakhaven', 'Millbrook', 'Riverdale', or similar generic compound settlement names. Every location should have a proper, distinctive name.
-- Before finalising, silently critique for: name originality, internal consistency (NPCs don't contradict each other), logical alignment between public face and secret agenda. Rewrite if issues found.`;
-
-        const userMessage = `Generate a faction. Variation seed: ${varianceSeed}.
-- Theme/Genre: ${theme}
-- Faction Type: ${factionType}
-- Scope: ${scope}
-- Moral Posture: ${alignment}${campaignContext ? `\n- Campaign Context: ${campaignContext}` : ""}
-- Faction Naming Directive: ${chosenNamingStyle}
-- NPC Naming Directive: ${chosenNpcStyle}`;
-
-        const model = await this.clientManager.getModel(
-          "",
-          "gemini-3.1-flash-lite",
-          systemInstruction,
-        );
-        const response = await model.generateContent(userMessage);
-        const text = response.response.text().trim();
-        const cleanText = text
-          .replace(/^```json\s*/i, "")
-          .replace(/```$/, "")
-          .trim();
-        const data = JSON.parse(cleanText);
-
-        return {
-          type: "faction",
-          title: data.title || name,
-          summary: data.summary || "",
-          content: data.content || "",
-          lore: data.lore || "",
-          labels: Array.isArray(data.labels)
-            ? data.labels
-            : ["rpg-faction", "faction-generator", "imported-draft"],
-          status: "active",
-        };
-      } catch (err) {
-        console.warn(
-          "AI generation failed, falling back to local tables:",
-          err,
-        );
-      }
-    }
-
-    const goal =
-      factionConfig.goals[
-        Math.floor(Math.random() * factionConfig.goals.length)
-      ];
-    const conflict =
-      factionConfig.conflicts[
-        Math.floor(Math.random() * factionConfig.conflicts.length)
-      ];
-    const hook =
-      factionConfig.hooks[
-        Math.floor(Math.random() * factionConfig.hooks.length)
-      ];
-    const rival = `${this.generateName()} Covenant`;
-    const leader = this.generateName();
-    const agent = this.generateName();
-
-    const summary = `A ${alignment.toLowerCase()} ${factionType.toLowerCase()} operating at the ${scope.toLowerCase()} level.`;
-
-    const content = `### What they control
-${name} is a ${factionType.toLowerCase()} with a firm grip on key resources across the ${scope.toLowerCase()}. Their reach is felt in every trade deal, guarded rumor, and carefully placed favor.${campaignContext ? ` In ${campaignContext}, they already have fingers in the most contested disputes.` : ""}
-
-### What they want
-${goal} Every action the faction takes, however charitable it appears, serves this underlying drive.
-
-### Why they are dangerous
-${conflict} Beyond their internal tensions, they will negotiate before striking — but they do not forget.
-
-### How to use them at the table
-Bring ${name} into play when the party needs leverage, pressure, a sponsor, or a rival who can operate in daylight. They reward players who deal in favors and punish those who make public enemies.`;
-
-    const lore = `### At the Table
-- **Base**: ${this.factionBase(factionType)}
-- **Resource**: ${this.factionResource(factionType)}
-- **Symbol**: ${name.split(" ")[0]} iconography worn by inner-circle members
-- **Secret**: ${conflict}
-- **Immediate hook**: ${hook}
-
-### Notable NPCs
-- **${leader}**: Public face who insists every deal serves the common good.
-- **${agent}**: Field operative who knows where the faction buries its failures.
-
-### Internal Conflict
-${conflict}
-
-### Rival Faction
-${rival} is pursuing the same influence, relic, or route — and will reach it first if the party does nothing.`;
-
-    return {
-      type: "faction",
-      title: name,
-      summary,
-      content,
-      lore,
-      labels: ["rpg-faction", "faction-generator", "imported-draft"],
-      status: "active",
-    };
+    return generateFaction(this.clientManager, options);
   }
 
-  /**
-   * Generates a vampire clan draft.
-   */
   async generateVampireClan(
-    options: {
-      archetype?: string;
-      bloodline?: string;
-      feedingHabit?: string;
-      weakness?: string;
-      scope?: string;
-      alignment?: string;
-      campaignContext?: string;
-      useAI?: boolean;
-    } = {},
+    options: Parameters<typeof generateVampireClan>[1] = {},
   ): Promise<GeneratorOutput> {
-    const archetype =
-      options.archetype ||
-      vampireConfig.archetypes[
-        Math.floor(Math.random() * vampireConfig.archetypes.length)
-      ];
-    const bloodline =
-      options.bloodline ||
-      vampireConfig.bloodlines[
-        Math.floor(Math.random() * vampireConfig.bloodlines.length)
-      ];
-    const feedingHabit =
-      options.feedingHabit ||
-      vampireConfig.feedingHabits[
-        Math.floor(Math.random() * vampireConfig.feedingHabits.length)
-      ];
-    const weakness =
-      options.weakness ||
-      vampireConfig.weaknesses[
-        Math.floor(Math.random() * vampireConfig.weaknesses.length)
-      ];
-    const scope =
-      options.scope ||
-      vampireConfig.scopes[
-        Math.floor(Math.random() * vampireConfig.scopes.length)
-      ];
-    const alignment =
-      options.alignment ||
-      vampireConfig.alignments[
-        Math.floor(Math.random() * vampireConfig.alignments.length)
-      ];
-    const campaignContext = options.campaignContext?.trim();
-
-    // Surnames/houses/covenants for vampires
-    const prefixes = ["House ", "The ", "Covenant of ", "Order of ", "Clan "];
-    const roots = [
-      "Dracul",
-      "Karnstein",
-      "Von Carstein",
-      "Orlok",
-      "Bathory",
-      "Tepes",
-      "Morbius",
-      "Sanguis",
-      "Vargo",
-      "Ruthven",
-    ];
-    const name =
-      prefixes[Math.floor(Math.random() * prefixes.length)] +
-      roots[Math.floor(Math.random() * roots.length)];
-
-    if (options.useAI !== false) {
-      try {
-        const prompt = `Generate a detailed RPG vampire clan/faction in JSON format.
-Options:
-- Name: ${name}
-- Clan Archetype: ${archetype}
-- Bloodline: ${bloodline}
-- Feeding Habit: ${feedingHabit}
-- Clan Weakness: ${weakness}
-- Scope of Influence: ${scope}
-- Moral Posture: ${alignment}
-${campaignContext ? `- Campaign Context: ${campaignContext}` : ""}
-
-You must return a valid JSON object matching the following structure exactly:
-{
-  "title": "A single string for the vampire clan/house name",
-  "content": "A detailed multi-paragraph overview (markdown formatted) describing its history, public facade in mortal society, dark haven, and how it fits the campaign context if provided.",
-  "lore": "Structured GM details (markdown formatted) with sections for core fields, bloodline traits, feeding habits, weakness, dark agenda, internal conflict, notable NPCs, rival faction, and adventure hook.",
-  "labels": ["rpg-faction", "vampire-clan", "imported-draft"]
-}
-Return only the JSON object. Do not include markdown code block formatting like \`\`\`json.`;
-
-        const model = await this.clientManager.getModel(
-          "",
-          "gemini-1.5-flash",
-          "You are an assistant that generates detailed RPG campaign elements in JSON format.",
-        );
-        const response = await model.generateContent(prompt);
-        const text = response.response.text().trim();
-        const cleanText = text
-          .replace(/^```json\s*/i, "")
-          .replace(/```$/, "")
-          .trim();
-        const data = JSON.parse(cleanText);
-
-        return {
-          type: "faction",
-          title: data.title || name,
-          content: data.content || "",
-          lore: data.lore || "",
-          labels: Array.isArray(data.labels)
-            ? data.labels
-            : ["rpg-faction", "vampire-clan", "imported-draft"],
-          status: "active",
-        };
-      } catch (err) {
-        console.warn(
-          "AI generation failed, falling back to local tables:",
-          err,
-        );
-      }
-    }
-
-    const goal =
-      vampireConfig.goals[
-        Math.floor(Math.random() * vampireConfig.goals.length)
-      ];
-    const conflict =
-      vampireConfig.conflicts[
-        Math.floor(Math.random() * vampireConfig.conflicts.length)
-      ];
-    const hook =
-      vampireConfig.hooks[
-        Math.floor(Math.random() * vampireConfig.hooks.length)
-      ];
-    const rival = `${this.generateName()} Inquisition`;
-    const sire = this.generateName();
-    const thrall = this.generateName();
-
-    const content = `### Overview
-${name} is a powerful vampire clan of the ${bloodline.toLowerCase()} lineage, operating as a ${archetype.toLowerCase()} across ${scope.toLowerCase()}. They hide their predatory activities behind a carefully crafted mortal facade, manipulating events from the dark.
-
-${campaignContext ? `### Campaign Fit\nUse ${name} in ${campaignContext}. Their influence should touch the local halls of power, forgotten catacombs, or ongoing dark mysteries.\n` : ""}
-
-### Public Facade
-To the mortal world, members of ${name} present themselves as wealthy philanthropists, eccentric scholars, or influential patrons. Very few suspect that behind this elegant mask lies a highly organized coven of undead hunters.
-
-### Table Use
-Introduce ${name} when the party enters high-society intrigue, investigates occult occurrences, or needs a powerful but dangerous ally who demands blood or secrets as currency.`;
-
-    const lore = `### GM Reference Information
-- **Faction Type**: Vampire Clan (${archetype})
-- **Bloodline**: ${bloodline}
-- **Scope**: ${scope}
-- **Moral Posture**: ${alignment}
-- **Feeding Habit**: ${feedingHabit}
-- **Clan Weakness**: ${weakness}
-- **Entity Type**: Faction
-
-### Dark Agenda
-${goal}
-
-### Internal Conflict
-${conflict}
-
-### Notable NPCs
-- **Sire ${sire}**: The ancient leader of the clan who has survived centuries of inquisitions and power struggles.
-- **Thrall ${thrall}**: A high-ranking mortal puppet who manages the clan's daytime assets and legal matters.
-
-### Rival Faction
-The ${rival} seeks to expose, purge, or take control of the secrets and assets held by ${name}.
-
-### Adventure Hook
-${hook}`;
-
-    return {
-      type: "faction",
-      title: name,
-      content,
-      lore,
-      labels: ["rpg-faction", "vampire-clan", "imported-draft"],
-      status: "active",
-    };
+    return generateVampireClan(this.clientManager, options);
   }
 
-  /**
-   * Generates a settlement draft.
-   */
   async generateSettlement(
-    options: {
-      size?: string;
-      economy?: string;
-      useAI?: boolean;
-    } = {},
+    options: Parameters<typeof generateSettlement>[1] = {},
   ): Promise<GeneratorOutput> {
-    const sizeConfig =
-      settlementConfig.sizes.find((s) => s.name === options.size) ||
-      settlementConfig.sizes[
-        Math.floor(Math.random() * settlementConfig.sizes.length)
-      ];
-    const size = sizeConfig.name;
-    const population = sizeConfig.range;
-    const economy =
-      options.economy ||
-      settlementConfig.economies[
-        Math.floor(Math.random() * settlementConfig.economies.length)
-      ];
-
-    // Generate Name using location styling
-    const prefixes = [
-      "Cinderwall",
-      "Stonebridge",
-      "Coppergate",
-      "Ashveil",
-      "Saltmarsh",
-      "Redthorn",
-      "Greywarden",
-      "Deepwell",
-    ];
-    const suffixes = [
-      " Crossing",
-      " Keep",
-      " Village",
-      " Town",
-      " Harbour",
-      " Hollow",
-      " Falls",
-      " Ridge",
-    ];
-    const name =
-      prefixes[Math.floor(Math.random() * prefixes.length)] +
-      suffixes[Math.floor(Math.random() * suffixes.length)];
-
-    if (options.useAI !== false) {
-      try {
-        const prompt = `Generate a detailed RPG Settlement in JSON format.
-Options:
-- Name: ${name}
-- Size: ${size} (${population})
-- Primary Economy: ${economy}
-
-You must return a valid JSON object matching the following structure exactly:
-{
-  "title": "A single string for the settlement name",
-  "content": "A detailed multi-paragraph description (markdown formatted) describing the settlement's atmosphere, layout, and geography.",
-  "lore": "Structured GM details (markdown formatted) detailing its size, government, notable locations, active factions, and current rumors.",
-  "labels": ["rpg-location", "imported-draft"]
-}
-Return only the JSON object. Do not include markdown code block formatting like \`\`\`json.`;
-
-        const model = await this.clientManager.getModel(
-          "",
-          "gemini-3.1-flash-lite",
-          "You are an assistant that generates detailed RPG campaign elements in JSON format.",
-        );
-        const response = await model.generateContent(prompt);
-        const text = response.response.text().trim();
-        const cleanText = text
-          .replace(/^```json\s*/i, "")
-          .replace(/```$/, "")
-          .trim();
-        const data = JSON.parse(cleanText);
-
-        return {
-          type: "location",
-          title: data.title || name,
-          content: data.content || "",
-          lore: data.lore || "",
-          labels: Array.isArray(data.labels)
-            ? data.labels
-            : ["rpg-location", "imported-draft"],
-          status: "active",
-        };
-      } catch (err) {
-        console.warn(
-          "AI generation failed, falling back to local tables:",
-          err,
-        );
-      }
-    }
-
-    // Local Fallback
-    const government =
-      settlementConfig.governments[
-        Math.floor(Math.random() * settlementConfig.governments.length)
-      ];
-    const faction =
-      settlementConfig.factions[
-        Math.floor(Math.random() * settlementConfig.factions.length)
-      ];
-    const locs = this.getRandomItems(
-      settlementConfig.notableLocations,
-      sizeConfig.pointsOfInterestCount,
-    );
-
-    const content = `### Description
-${name} is a thriving ${size.toLowerCase()} situated along major geography channels. Its local architecture features sturdy foundations tailored to the environment, and its streets are active with citizens going about their daily routines.
-
-### Atmosphere
-The air is filled with the smells of local industries, and visitors are greeted with a blend of curiosity and wariness depending on the active hour. The economy is heavily focused on ${economy.toLowerCase()} operations, serving as the main source of income for local families.`;
-
-    const lore = `### GM Reference Information
-- **Size**: ${size} (${population})
-- **Primary Economy**: ${economy}
-- **Government**: ${government}
-
-### Points of Interest
-${locs.map((l) => `- **${l}**: A crucial hub of local activity.`).join("\n")}
-
-### Controlling Factions
-- **${faction}**: Maintains significant influence over the local district's rules and affairs.`;
-
-    return {
-      type: "location",
-      title: name,
-      content,
-      lore,
-      labels: ["rpg-location", "imported-draft"],
-      status: "active",
-    };
+    return generateSettlement(this.clientManager, options);
   }
 
-  /**
-   * Generates a magic item draft.
-   */
   async generateMagicItem(
-    options: {
-      type?: string;
-      rarity?: string;
-      useAI?: boolean;
-    } = {},
+    options: Parameters<typeof generateMagicItem>[1] = {},
   ): Promise<GeneratorOutput> {
-    const itemType =
-      options.type ||
-      magicItemConfig.types[
-        Math.floor(Math.random() * magicItemConfig.types.length)
-      ];
-    const rarity =
-      options.rarity ||
-      magicItemConfig.rarities[
-        Math.floor(Math.random() * magicItemConfig.rarities.length)
-      ];
-
-    // Generate Name using descriptors
-    const prefixes = [
-      "Dread",
-      "Aether",
-      "Frost",
-      "Shadow",
-      "Soul",
-      "Solar",
-      "Storm",
-      "Whisper",
-      "Rune",
-    ];
-    const suffixes = [
-      "bringer",
-      "weaver",
-      "ward",
-      "shard",
-      "reaper",
-      "binder",
-      "heart",
-      "caller",
-    ];
-    const baseName =
-      prefixes[Math.floor(Math.random() * prefixes.length)] +
-      suffixes[Math.floor(Math.random() * suffixes.length)];
-    const name = `${baseName} (${itemType})`;
-
-    if (options.useAI !== false) {
-      try {
-        const prompt = `Generate a detailed RPG Magic Item in JSON format.
-Options:
-- Name: ${name}
-- Type: ${itemType}
-- Rarity: ${rarity}
-
-You must return a valid JSON object matching the following structure exactly:
-{
-  "title": "A single string for the magic item name",
-  "content": "A detailed description (markdown formatted) describing the item's appearance, materials, and passive feelings when held.",
-  "lore": "Structured GM details (markdown formatted) detailing its magical properties, rarity, curse (if any), and legendary backstory.",
-  "labels": ["rpg-item", "imported-draft"]
-}
-Return only the JSON object. Do not include markdown code block formatting like \`\`\`json.`;
-
-        const model = await this.clientManager.getModel(
-          "",
-          "gemini-3.1-flash-lite",
-          "You are an assistant that generates detailed RPG campaign elements in JSON format.",
-        );
-        const response = await model.generateContent(prompt);
-        const text = response.response.text().trim();
-        const cleanText = text
-          .replace(/^```json\s*/i, "")
-          .replace(/```$/, "")
-          .trim();
-        const data = JSON.parse(cleanText);
-
-        return {
-          type: "item",
-          title: data.title || name,
-          content: data.content || "",
-          lore: data.lore || "",
-          labels: Array.isArray(data.labels)
-            ? data.labels
-            : ["rpg-item", "imported-draft"],
-          status: "active",
-        };
-      } catch (err) {
-        console.warn(
-          "AI generation failed, falling back to local tables:",
-          err,
-        );
-      }
-    }
-
-    // Local Fallback
-    const property =
-      magicItemConfig.properties[
-        Math.floor(Math.random() * magicItemConfig.properties.length)
-      ];
-    const history =
-      magicItemConfig.histories[
-        Math.floor(Math.random() * magicItemConfig.histories.length)
-      ];
-
-    const content = `### Description
-The ${name} is a uniquely crafted ${itemType.toLowerCase()} that displays a high degree of precision in its construction. Made from materials rare to this region, it feels slightly warm or cool to the touch depending on the active wielder's alignment.`;
-
-    const lore = `### GM Reference Information
-- **Type**: ${itemType}
-- **Rarity**: ${rarity}
-
-### Magical Properties
-- **Passive Effect**: ${property}
-
-### Lore & History
-${history}`;
-
-    return {
-      type: "item",
-      title: name,
-      content,
-      lore,
-      labels: ["rpg-item", "imported-draft"],
-      status: "active",
-    };
+    return generateMagicItem(this.clientManager, options);
   }
 
   async generateQuestHook(
-    options: {
-      genre?: string;
-      tone?: string;
-      scope?: string;
-      locationType?: string;
-      threat?: string;
-      twist?: string;
-      reward?: string;
-      campaignContext?: string;
-      useAI?: boolean;
-    } = {},
+    options: Parameters<typeof generateQuestHook>[1] = {},
   ): Promise<GeneratorOutput> {
-    const genre =
-      options.genre ||
-      questConfig.genres[Math.floor(Math.random() * questConfig.genres.length)];
-    const tone =
-      options.tone ||
-      questConfig.tones[Math.floor(Math.random() * questConfig.tones.length)];
-    const scope =
-      options.scope ||
-      questConfig.scopes[Math.floor(Math.random() * questConfig.scopes.length)];
-    const locationType =
-      options.locationType ||
-      questConfig.locationTypes[
-        Math.floor(Math.random() * questConfig.locationTypes.length)
-      ];
-    const threat =
-      options.threat ||
-      questConfig.threats[
-        Math.floor(Math.random() * questConfig.threats.length)
-      ];
-    const twist =
-      options.twist ||
-      questConfig.twists[Math.floor(Math.random() * questConfig.twists.length)];
-    const reward =
-      options.reward ||
-      questConfig.rewards[
-        Math.floor(Math.random() * questConfig.rewards.length)
-      ];
-    const campaignContext = options.campaignContext?.trim();
-    const questName = `${this.generateName()}'s ${["Gambit", "Bargain", "Reckoning", "Shadow", "Legacy", "Trial"][Math.floor(Math.random() * 6)]}`;
-
-    if (options.useAI !== false) {
-      try {
-        const prompt = `Generate a detailed RPG quest hook in JSON format.
-Options:
-- Genre: ${genre}
-- Tone: ${tone}
-- Scope: ${scope}
-- Location Type: ${locationType}
-- Main Threat: ${threat}
-- Twist: ${twist}
-- Reward: ${reward}
-${campaignContext ? `- Campaign Context: ${campaignContext}` : ""}
-
-You must return a valid JSON object matching the following structure exactly:
-{
-  "title": "A single evocative quest name (3-6 words)",
-  "content": "A detailed multi-paragraph quest hook (markdown formatted) describing the situation, what the party is asked to do, the location, the key NPC involved, and how it fits the campaign context if provided.",
-  "lore": "Structured GM details (markdown formatted) with sections for core fields, complication, key NPC, twist, and reward.",
-  "labels": ["rpg-quest", "quest-generator", "imported-draft"]
-}
-Return only the JSON object. Do not include markdown code block formatting like \`\`\`json.`;
-
-        const model = await this.clientManager.getModel(
-          "",
-          "gemini-3.1-flash-lite",
-          "You are an assistant that generates detailed RPG campaign elements in JSON format.",
-        );
-        const response = await model.generateContent(prompt);
-        const text = response.response.text().trim();
-        const cleanText = text
-          .replace(/^```json\s*/i, "")
-          .replace(/```$/, "")
-          .trim();
-        const data = JSON.parse(cleanText);
-
-        return {
-          type: "event",
-          title: data.title || questName,
-          content: data.content || "",
-          lore: data.lore || "",
-          labels: Array.isArray(data.labels)
-            ? data.labels
-            : ["rpg-quest", "quest-generator", "imported-draft"],
-          status: "active",
-        };
-      } catch (err) {
-        console.warn(
-          "AI generation failed, falling back to local tables:",
-          err,
-        );
-      }
-    }
-
-    const hook =
-      questConfig.hooks[Math.floor(Math.random() * questConfig.hooks.length)];
-    const complication =
-      questConfig.complications[
-        Math.floor(Math.random() * questConfig.complications.length)
-      ];
-    const npcName = this.generateName();
-    const locationName = `The ${this.generateName()} ${locationType}`;
-
-    const content = `### The Hook
-${hook}
-
-${campaignContext ? `### Campaign Fit\nThis quest ties into ${campaignContext}. The threat and location should reflect existing tensions or unresolved threads.\n` : ""}### Location
-${locationName} serves as the primary setting — a ${locationType.toLowerCase()} shaped by ${genre.toLowerCase()} conventions and a ${tone.toLowerCase()} atmosphere.
-
-### Key NPC
-**${npcName}** is the immediate contact, patron, or obstacle. Their stated reason for hiring the party is credible enough, but their personal stake runs deeper than they admit.
-
-### Threat
-The central danger is a ${threat.toLowerCase()}. It has been active long enough to leave evidence, earn fear, and create a power vacuum that others are already trying to fill.`;
-
-    const lore = `### GM Reference Information
-- **Genre**: ${genre}
-- **Tone**: ${tone}
-- **Scope**: ${scope}
-- **Location Type**: ${locationType}
-- **Main Threat**: ${threat}
-
-### Complication
-${complication}
-
-### Twist
-${twist}
-
-### Reward
-${reward}`;
-
-    return {
-      type: "event",
-      title: questName,
-      content,
-      lore,
-      labels: ["rpg-quest", "quest-generator", "imported-draft"],
-      status: "active",
-    };
+    return generateQuestHook(this.clientManager, options);
   }
 
   async generateNames(
-    options: {
-      culture?: string;
-      gender?: string;
-      nameType?: string;
-      count?: string;
-      context?: string;
-      useAI?: boolean;
-    } = {},
+    options: Parameters<typeof generateNames>[1] = {},
   ): Promise<GeneratorOutput> {
-    const culture = options.culture || nameGeneratorConfig.cultures[0];
-    const gender = options.gender || nameGeneratorConfig.genders[0];
-    const nameType = options.nameType || nameGeneratorConfig.nameTypes[0];
-    const count = Math.max(1, parseInt(options.count || "5", 10) || 5);
-    const context = options.context?.trim();
-
-    const entityType: GeneratorOutput["type"] =
-      nameType === "Place"
-        ? "location"
-        : nameType === "Faction"
-          ? "faction"
-          : nameType === "Item"
-            ? "item"
-            : "character";
-
-    if (options.useAI !== false) {
-      try {
-        const prompt = `Generate ${count} fantasy ${nameType.toLowerCase()} names in JSON format.
-Options:
-- Culture / Style: ${culture}
-- Gender / Presentation: ${gender}
-- Name Type: ${nameType}
-- Count: ${count}
-${context ? `- Context: ${context}` : ""}
-
-You must return a valid JSON object matching the following structure exactly:
-{
-  "title": "The single best or most evocative name from the list",
-  "content": "A brief lead sentence describing the naming style, followed by a markdown list of all ${count} names. Format each name as '- **Name**: one-sentence flavour note'.",
-  "lore": "GM notes (markdown formatted) with sections for Culture, Style, and Usage Suggestions covering how to use these names in a campaign.",
-  "labels": ["fantasy-name", "name-generator", "imported-draft"]
-}
-Return only the JSON object. Do not include markdown code block formatting like \`\`\`json.`;
-
-        const model = await this.clientManager.getModel(
-          "",
-          "gemini-3.1-flash-lite",
-          "You are an assistant that generates detailed RPG campaign elements in JSON format.",
-        );
-        const response = await model.generateContent(prompt);
-        const text = response.response.text().trim();
-        const cleanText = text
-          .replace(/^```json\s*/i, "")
-          .replace(/```$/, "")
-          .trim();
-        const data = JSON.parse(cleanText);
-
-        return {
-          type: entityType,
-          title: data.title || "Fantasy Name",
-          content: data.content || "",
-          lore: data.lore || "",
-          labels: Array.isArray(data.labels)
-            ? data.labels
-            : ["fantasy-name", "name-generator", "imported-draft"],
-          status: "active",
-        };
-      } catch (err) {
-        console.warn(
-          "AI generation failed, falling back to local tables:",
-          err,
-        );
-      }
-    }
-
-    // Local fallback: combine culture-specific prefix + suffix
-    const prefixes =
-      nameGeneratorConfig.culturePrefixes[culture] ||
-      nameGeneratorConfig.culturePrefixes["Generic Fantasy"];
-    const suffixes =
-      nameGeneratorConfig.cultureSuffixes[culture] ||
-      nameGeneratorConfig.cultureSuffixes["Generic Fantasy"];
-
-    const generated: string[] = [];
-    for (let i = 0; i < count; i++) {
-      const prefix = prefixes[Math.floor(Math.random() * prefixes.length)];
-      const suffix = suffixes[Math.floor(Math.random() * suffixes.length)];
-      generated.push(prefix.charAt(0).toUpperCase() + prefix.slice(1) + suffix);
-    }
-
-    const primary = generated[0];
-    const nameList = generated.map((n) => `- ${n}`).join("\n");
-
-    const content = `${culture} ${nameType.toLowerCase()} names in a ${gender.toLowerCase()} register.
-
-${nameList}`;
-
-    const lore = `### Generator Settings
-- **Culture**: ${culture}
-- **Gender / Presentation**: ${gender}
-- **Name Type**: ${nameType}
-
-### Usage Suggestions
-Use these names for any ${nameType.toLowerCase()} in a ${culture.toLowerCase()}-influenced setting. Combine or modify them freely — drop a syllable, add a prefix, or append a title or epithet for variation.`;
-
-    return {
-      type: entityType,
-      title: primary,
-      content,
-      lore,
-      labels: ["fantasy-name", "name-generator", "imported-draft"],
-      status: "active",
-    };
+    return generateNames(this.clientManager, options);
   }
 
   async generateSocialHub(
-    options: {
-      genre?: string;
-      venueType?: string;
-      atmosphere?: string;
-      wealthLevel?: string;
-      clientele?: string;
-      campaignContext?: string;
-      useAI?: boolean;
-    } = {},
+    options: Parameters<typeof generateSocialHub>[1] = {},
   ): Promise<GeneratorOutput> {
-    const genre =
-      options.genre ||
-      socialHubConfig.genres[
-        Math.floor(Math.random() * socialHubConfig.genres.length)
-      ];
-    const venueTypes =
-      socialHubConfig.venueTypesByGenre[genre] ??
-      socialHubConfig.venueTypesByGenre["Fantasy"];
-    const venueType =
-      options.venueType ||
-      venueTypes[Math.floor(Math.random() * venueTypes.length)];
-    const atmosphere =
-      options.atmosphere ||
-      socialHubConfig.atmospheres[
-        Math.floor(Math.random() * socialHubConfig.atmospheres.length)
-      ];
-    const wealthLevel = options.wealthLevel || socialHubConfig.wealthLevels[2];
-    const clienteles =
-      socialHubConfig.clientelesByGenre[genre] ??
-      socialHubConfig.clientelesByGenre["Fantasy"];
-    const clientele =
-      options.clientele ||
-      clienteles[Math.floor(Math.random() * clienteles.length)];
-    const campaignContext = options.campaignContext?.trim();
-
-    const varianceSeed = Math.floor(Math.random() * 99991) + 10;
-
-    if (options.useAI !== false) {
-      try {
-        const systemInstruction = `You are an expert RPG campaign writer. You generate immediately usable social gathering locations for tabletop GMs in JSON format. You write in the register of the specified genre — a cyberpunk noodle bar sounds nothing like a fantasy tavern.
-
-OUTPUT FORMAT — return ONLY a valid JSON object, no markdown fences:
-{
-  "title": "The venue's name — genre-appropriate (a cyberpunk dive has a neon handle, a fantasy inn has 'The X Y' naming, a western saloon uses frontier naming, etc.)",
-  "summary": "One sentence: the venue's character and why a GM should use it.",
-  "content": "Markdown. Use exactly these four section headers in order: '### The Place', '### The People', '### The Trouble', '### How to use it at the table'. Each section: 2-4 tight sentences. Write in the genre's register. Include campaign context if provided.",
-  "lore": "Markdown. Use EXACTLY this structure:\\n### At a Glance\\n- **Type**: venue type\\n- **Atmosphere**: mood and feel\\n- **Owner / Operator**: name and one-line description (genre-appropriate)\\n- **Signature Drink or Service**: specific invented item or service\\n- **Hidden Problem**: the trouble simmering beneath the surface\\n- **Immediate Hook**: one-sentence GM hook\\n### Notable Regulars\\n- **Name**: one-line description (2-3 regulars, genre-appropriate)\\n### Rumours\\n- short rumour (2-3 rumours as bullet points)\\n### Entity Seeds\\n- list of 3-4 Codex entity types that naturally emerge from this venue (e.g. '**Location**: The back room')",
-  "labels": ["2-4 lowercase genre-appropriate tags, plus 'rpg-location', 'social-hub-generator', 'imported-draft'"]
-}
-
-QUALITY RULES:
-- Everything — names, slang, concerns, technology — must fit the genre.
-- The owner/operator must have a name, a distinguishing detail, and a secret or motive.
-- Each regular should feel like a distinct person with a reason to be there.
-- Rumours should be specific, not generic.
-- Entity seeds should suggest concrete Codex entries a GM could create.`;
-
-        const userMessage = `Generate a social gathering location. Variation seed: ${varianceSeed}.
-- Genre / Setting: ${genre}
-- Venue Type: ${venueType}
-- Atmosphere: ${atmosphere}
-- Wealth Level: ${wealthLevel}
-- Primary Clientele: ${clientele}${campaignContext ? `\n- Campaign Context: ${campaignContext}` : ""}`;
-
-        const model = await this.clientManager.getModel(
-          "",
-          "gemini-3.1-flash-lite",
-          systemInstruction,
-        );
-        const response = await model.generateContent(userMessage);
-        const text = response.response.text().trim();
-        const cleanText = text
-          .replace(/^```json\s*/i, "")
-          .replace(/```$/, "")
-          .trim();
-        const data = JSON.parse(cleanText);
-
-        return {
-          type: "location",
-          title: data.title || "The Unnamed Venue",
-          summary: data.summary || "",
-          content: data.content || "",
-          lore: data.lore || "",
-          labels: Array.isArray(data.labels)
-            ? data.labels
-            : ["rpg-location", "social-hub-generator", "imported-draft"],
-          status: "active",
-        };
-      } catch (err) {
-        console.warn(
-          "AI generation failed, falling back to local tables:",
-          err,
-        );
-      }
-    }
-
-    // Local fallback
-    const trouble =
-      socialHubConfig.troubles[
-        Math.floor(Math.random() * socialHubConfig.troubles.length)
-      ];
-    const ownerName = this.generateName();
-    const patron1 = this.generateName();
-    const patron2 = this.generateName();
-    const rawName = this.generateName().replace(/\s+The\s+\S+$/, "");
-
-    const fantasyLikeSuffixes = ["Arms", "Rest", "Tap", "Lodge", "House"];
-    const cyberpunkSuffixes = ["Node", "Den", "Hub", "Spot", "Joint"];
-    const westernSuffixes = ["Saloon", "House", "Post", "Room", "Bar"];
-    const suffixMap: Record<string, string[]> = {
-      Fantasy: fantasyLikeSuffixes,
-      "Dark Fantasy": fantasyLikeSuffixes,
-      Pirate: ["Cove", "Hole", "Deck", "Anchor", "Port"],
-      Cyberpunk: cyberpunkSuffixes,
-      "Sci-Fi": ["Bay", "Dock", "Zone", "Platform", "Hub"],
-      Modern: ["Bar", "Lounge", "Spot", "Place", "Corner"],
-      Horror: ["Den", "Hollow", "Parlour", "Chamber", "Haunt"],
-      "Post-Apocalyptic": ["Shack", "Hole", "Stop", "Den", "Post"],
-      Western: westernSuffixes,
-    };
-    const suffixes = suffixMap[genre] ?? fantasyLikeSuffixes;
-    const venueName =
-      genre === "Cyberpunk" || genre === "Sci-Fi"
-        ? `${rawName}'s ${suffixes[Math.floor(Math.random() * suffixes.length)]}`
-        : `The ${rawName} ${suffixes[Math.floor(Math.random() * suffixes.length)]}`;
-
-    const summary = `A ${atmosphere.toLowerCase()} ${venueType.toLowerCase()} serving ${clientele.toLowerCase()}.`;
-
-    const content = `### The Place
-${venueName} is a ${venueType.toLowerCase()} catering to ${clientele.toLowerCase()}. The atmosphere is ${atmosphere.toLowerCase()}, and the wealth level is ${wealthLevel.toLowerCase()}.${campaignContext ? ` In ${campaignContext}, it sits at the edge of the main conflict.` : ""}
-
-### The People
-The operator, ${ownerName}, runs a tight establishment. Regular faces include ${patron1}, a well-known local, and ${patron2}, who rarely speaks about where they sleep.
-
-### The Trouble
-${trouble}. The operator is aware of enough to be nervous, but not enough to act.
-
-### How to use it at the table
-Use ${venueName} as a home base, an information hub, or a pressure point. The trouble beneath the surface gives any visit the potential to escalate.`;
-
-    const lore = `### At a Glance
-- **Type**: ${venueType}
-- **Atmosphere**: ${atmosphere}
-- **Owner / Operator**: ${ownerName} — competent, guarded, and owed favours by the wrong people
-- **Signature Drink or Service**: The house special, served without questions
-- **Hidden Problem**: ${trouble}
-- **Immediate Hook**: A regular has not shown up for three days — their usual spot is still being held
-
-### Notable Regulars
-- **${patron1}**: A reliable local who knows more than they let on
-- **${patron2}**: A recent arrival who pays in currency that raises questions
-
-### Rumours
-- Someone was seen leaving through the back way well after hours
-- A back room was recently sealed off — the operator says maintenance, locals say otherwise
-- A faction has been asking after a specific regular who may have passed through recently
-
-### Entity Seeds
-- **Location**: ${venueName} (this venue)
-- **Character**: ${ownerName} (operator)
-- **Character**: ${patron1} (regular)
-- **Faction**: Whoever is behind the hidden trouble`;
-
-    return {
-      type: "location",
-      title: venueName,
-      summary,
-      content,
-      lore,
-      labels: ["rpg-location", "social-hub-generator", "imported-draft"],
-      status: "active",
-    };
+    return generateSocialHub(this.clientManager, options);
   }
 
   async generateTavern(
-    options: {
-      type?: string;
-      atmosphere?: string;
-      settlementType?: string;
-      wealthLevel?: string;
-      clientele?: string;
-      campaignContext?: string;
-      useAI?: boolean;
-    } = {},
+    options: Parameters<typeof generateTavern>[1] = {},
   ): Promise<GeneratorOutput> {
-    const tavernType =
-      options.type ||
-      socialHubConfig.venueTypesByGenre["Fantasy"][
-        Math.floor(
-          Math.random() * socialHubConfig.venueTypesByGenre["Fantasy"].length,
-        )
-      ];
-    const atmosphere =
-      options.atmosphere ||
-      socialHubConfig.atmospheres[
-        Math.floor(Math.random() * socialHubConfig.atmospheres.length)
-      ];
-    const settlementType =
-      options.settlementType ||
-      socialHubConfig.settlementTypes[
-        Math.floor(Math.random() * socialHubConfig.settlementTypes.length)
-      ];
-    const wealthLevel = options.wealthLevel || socialHubConfig.wealthLevels[2];
-    const clientele =
-      options.clientele ||
-      socialHubConfig.clientelesByGenre["Fantasy"][
-        Math.floor(
-          Math.random() * socialHubConfig.clientelesByGenre["Fantasy"].length,
-        )
-      ];
-    const campaignContext = options.campaignContext?.trim();
-
-    const namingStyles = [
-      "Name it after an animal and a colour or material (e.g. 'The Copper Boar').",
-      "Name it after an object associated with the owner's past — a weapon, trade tool, or keepsake.",
-      "Use a short ironic phrase that locals find funny (e.g. 'The Honest Scales').",
-      "Name it after a local legend, battle, or geographical feature specific to the setting.",
-      "Use a two-word compound that evokes the atmosphere — one noun, one adjective (e.g. 'The Sullen Lantern').",
-    ];
-    const varianceSeed = Math.floor(Math.random() * 99991) + 10;
-
-    if (options.useAI !== false) {
-      try {
-        const systemInstruction = `You are an expert RPG campaign writer. You generate immediately usable tavern and inn locations for tabletop GMs in JSON format.
-
-OUTPUT FORMAT — return ONLY a valid JSON object, no markdown fences:
-{
-  "title": "The tavern's name (follow the naming directive in the user message)",
-  "summary": "One sentence: the tavern's character and why a GM should use it (e.g. 'A cramped dockside alehouse where smugglers and off-duty guards share the same bad wine.').",
-  "content": "Markdown. Use exactly these four section headers in order: '### The Place', '### The People', '### The Trouble', '### How to use it at the table'. Each section: 2-4 tight sentences. Include campaign context if provided.",
-  "lore": "Markdown. Use EXACTLY this structure:\\n### At a Glance\\n- **Type**: tavern type\\n- **Atmosphere**: mood and feel\\n- **Owner**: name and one-line description\\n- **Signature Drink or Dish**: specific invented item\\n- **Hidden Problem**: the trouble simmering beneath the surface\\n- **Immediate Hook**: one-sentence GM hook\\n### Notable Patrons\\n- **Name**: one-line description (2-3 patrons)\\n### Rumours\\n- short rumour (2-3 rumours as bullet points)\\n### Entity Seeds\\n- list of 3-4 Codex entity types that naturally emerge from this tavern (e.g. '**Location**: The cellar passage')",
-  "labels": ["2-4 lowercase tags, plus 'rpg-location', 'tavern-generator', 'imported-draft'"]
-}
-
-QUALITY RULES:
-- The owner must have a name, a distinguishing physical detail, and a secret or motive.
-- Each patron should feel like a distinct person with a reason to be there.
-- Rumours should be specific, not generic ('a merchant was poisoned last week' beats 'strange things have been happening').
-- Entity seeds should suggest concrete Codex entries a GM could create.
-- Place names and NPC names must be invented — avoid Oakhaven, Millbrook, Riverdale, and generic monosyllable English surnames.`;
-
-        const userMessage = `Generate a tavern. Variation seed: ${varianceSeed}.
-- Type: ${tavernType}
-- Atmosphere: ${atmosphere}
-- Settlement Type: ${settlementType}
-- Wealth Level: ${wealthLevel}
-- Primary Clientele: ${clientele}${campaignContext ? `\n- Campaign Context: ${campaignContext}` : ""}
-- Naming Directive: ${namingStyles[Math.floor(Math.random() * namingStyles.length)]}`;
-
-        const model = await this.clientManager.getModel(
-          "",
-          "gemini-3.1-flash-lite",
-          systemInstruction,
-        );
-        const response = await model.generateContent(userMessage);
-        const text = response.response.text().trim();
-        const cleanText = text
-          .replace(/^```json\s*/i, "")
-          .replace(/```$/, "")
-          .trim();
-        const data = JSON.parse(cleanText);
-
-        return {
-          type: "location",
-          title: data.title || "The Unnamed Tavern",
-          summary: data.summary || "",
-          content: data.content || "",
-          lore: data.lore || "",
-          labels: Array.isArray(data.labels)
-            ? data.labels
-            : ["rpg-location", "tavern-generator", "imported-draft"],
-          status: "active",
-        };
-      } catch (err) {
-        console.warn(
-          "AI generation failed, falling back to local tables:",
-          err,
-        );
-      }
-    }
-
-    // Local fallback
-    const trouble =
-      socialHubConfig.troubles[
-        Math.floor(Math.random() * socialHubConfig.troubles.length)
-      ];
-    const ownerName = this.generateName();
-    const patron1 = this.generateName();
-    const patron2 = this.generateName();
-    const rawName = this.generateName().replace(/\s+The\s+\S+$/, "");
-    const tavernName = `The ${rawName} ${["Arms", "Rest", "Tap", "Lodge", "House"][Math.floor(Math.random() * 5)]}`;
-
-    const summary = `A ${atmosphere.toLowerCase()} ${tavernType.toLowerCase()} serving ${clientele.toLowerCase()} in a ${settlementType.toLowerCase()}.`;
-
-    const content = `### The Place
-${tavernName} is a ${tavernType.toLowerCase()} in a ${settlementType.toLowerCase()}, catering to ${clientele.toLowerCase()}. The atmosphere is ${atmosphere.toLowerCase()}, and the wealth level is ${wealthLevel.toLowerCase()}.${campaignContext ? ` In ${campaignContext}, it sits at the edge of the main conflict.` : ""}
-
-### The People
-The owner, ${ownerName}, runs a tight establishment. Regular patrons include ${patron1}, a well-known local face, and ${patron2}, who rarely speaks about where they sleep.
-
-### The Trouble
-${trouble}. The owner is aware of enough to be nervous, but not enough to act.
-
-### How to use it at the table
-Use ${tavernName} as a home base, a rumour hub, or a pressure point. The trouble beneath the surface gives any visit the potential to escalate.`;
-
-    const lore = `### At a Glance
-- **Type**: ${tavernType}
-- **Atmosphere**: ${atmosphere}
-- **Owner**: ${ownerName} — competent, guarded, and owed favours by the wrong people
-- **Signature Drink**: House ale brewed in the cellar, served warm
-- **Hidden Problem**: ${trouble}
-- **Immediate Hook**: A regular patron has not shown up for three days — their usual table is still reserved
-
-### Notable Patrons
-- **${patron1}**: A reliable local who knows more than they let on
-- **${patron2}**: A recent arrival who pays in coin that smells of somewhere far away
-
-### Rumours
-- Someone was seen leaving through the back door well after midnight
-- The cellar was recently bricked up — the owner says rats, locals say otherwise
-- A faction has been asking after a specific traveller who may have stayed here recently
-
-### Entity Seeds
-- **Location**: ${tavernName} (this tavern)
-- **Character**: ${ownerName} (owner)
-- **Character**: ${patron1} (regular patron)
-- **Faction**: Whoever is behind the hidden trouble`;
-
-    return {
-      type: "location",
-      title: tavernName,
-      summary,
-      content,
-      lore,
-      labels: ["rpg-location", "tavern-generator", "imported-draft"],
-      status: "active",
-    };
+    return generateTavern(this.clientManager, options);
   }
 
   async generateKingdom(
-    options: {
-      polityType?: string;
-      governmentStyle?: string;
-      geography?: string;
-      scale?: string;
-      conflictLevel?: string;
-      magicLevel?: string;
-      campaignContext?: string;
-      useAI?: boolean;
-    } = {},
+    options: Parameters<typeof generateKingdom>[1] = {},
   ): Promise<GeneratorOutput> {
-    const polityType =
-      options.polityType ||
-      kingdomConfig.polityTypes[
-        Math.floor(Math.random() * kingdomConfig.polityTypes.length)
-      ];
-    const governmentStyle =
-      options.governmentStyle ||
-      kingdomConfig.governmentStyles[
-        Math.floor(Math.random() * kingdomConfig.governmentStyles.length)
-      ];
-    const geography =
-      options.geography ||
-      kingdomConfig.geographies[
-        Math.floor(Math.random() * kingdomConfig.geographies.length)
-      ];
-    const scale = options.scale || kingdomConfig.scales[2];
-    const conflictLevel =
-      options.conflictLevel ||
-      kingdomConfig.conflictLevels[
-        Math.floor(Math.random() * kingdomConfig.conflictLevels.length)
-      ];
-    const magicLevel = options.magicLevel || kingdomConfig.magicLevels[2];
-    const campaignContext = options.campaignContext?.trim();
-    const varianceSeed = Math.floor(Math.random() * 99991) + 10;
-
-    if (options.useAI !== false) {
-      try {
-        const systemInstruction = `You are an expert RPG campaign writer. You generate immediately usable fantasy kingdoms and political entities for tabletop GMs in JSON format.
-
-OUTPUT FORMAT — return ONLY a valid JSON object, no markdown fences:
-{
-  "title": "The realm's name — invented, fantasy-appropriate, not a generic placeholder",
-  "summary": "One sentence: the kingdom's defining character and why a GM should use it.",
-  "content": "Markdown. Use exactly these four section headers in order: '### The Realm', '### Government & Power', '### Society & Culture', '### How to use it at the table'. Each section: 3-5 tight sentences. Include campaign context if provided.",
-  "lore": "Markdown. Use EXACTLY this structure:\\n### At a Glance\\n- **Polity Type**: ${polityType}\\n- **Ruler**: invented name and one-line description\\n- **Capital**: invented name and one-line description\\n- **Scale**: population/territory summary\\n- **Magic Level**: ${magicLevel}\\n- **Conflict Level**: ${conflictLevel}\\n- **Hidden Problem**: the tension simmering beneath the surface\\n- **Immediate Hook**: one-sentence GM hook\\n### Major Factions\\n- **Name**: one-line description (2-3 factions)\\n### Rumours & Hooks\\n- short hook (2-3 bullet points)\\n### Entity Seeds\\n- list of 4-5 Codex entity types (e.g. '**Character**: The regent', '**Faction**: The merchant guild', '**Location**: The capital city')",
-  "labels": ["2-4 lowercase tags, plus 'rpg-kingdom', 'kingdom-generator', 'imported-draft'"]
-}
-
-QUALITY RULES:
-- The ruler must have a name, a defining trait, and a secret motive.
-- Each faction should have a clear agenda that creates tension with another faction.
-- Rumours should be specific and actionable — something a party could investigate.
-- Entity seeds should suggest concrete Codex entries a GM would actually create.
-- All names must be invented — avoid generic English words as names.`;
-
-        const userMessage = `Generate a fantasy kingdom. Variation seed: ${varianceSeed}.
-- Polity Type: ${polityType}
-- Government Style: ${governmentStyle}
-- Geography: ${geography}
-- Scale: ${scale}
-- Conflict Level: ${conflictLevel}
-- Magic Level: ${magicLevel}${campaignContext ? `\n- Campaign Context: ${campaignContext}` : ""}`;
-
-        const model = await this.clientManager.getModel(
-          "",
-          "gemini-3.1-flash-lite",
-          systemInstruction,
-        );
-        const response = await model.generateContent(userMessage);
-        const text = response.response.text().trim();
-        const cleanText = text
-          .replace(/^```json\s*/i, "")
-          .replace(/```$/, "")
-          .trim();
-        const data = JSON.parse(cleanText);
-
-        return {
-          type: "faction",
-          title: data.title || "The Unnamed Realm",
-          summary: data.summary || "",
-          content: data.content || "",
-          lore: data.lore || "",
-          labels: Array.isArray(data.labels)
-            ? data.labels
-            : ["rpg-kingdom", "kingdom-generator", "imported-draft"],
-          status: "active",
-        };
-      } catch (err) {
-        console.warn(
-          "AI generation failed, falling back to local tables:",
-          err,
-        );
-      }
-    }
-
-    // Local fallback
-    const trouble =
-      kingdomConfig.troubles[
-        Math.floor(Math.random() * kingdomConfig.troubles.length)
-      ];
-    const rulerName = this.generateName();
-    const factionName1 = `The ${this.generateName()} ${["Order", "Council", "League", "Brotherhood", "Guild"][Math.floor(Math.random() * 5)]}`;
-    const factionName2 = `The ${this.generateName()} ${["Compact", "Circle", "Assembly", "Society", "House"][Math.floor(Math.random() * 5)]}`;
-    const rawName = this.generateName().replace(/\s+The\s+\S+$/, "");
-    const realmName = `The ${["Kingdom", "Realm", "Domain", "Dominion", "Principality"][Math.floor(Math.random() * 5)]} of ${rawName}`;
-    const capitalName = this.generateName().replace(/\s+The\s+\S+$/, "");
-
-    const summary = `A ${conflictLevel.toLowerCase()} ${polityType.toLowerCase()} set in ${geography.toLowerCase()}, ruled by a ${governmentStyle.toLowerCase()}.`;
-
-    const content = `### The Realm
-${realmName} is a ${scale.toLowerCase()} ${polityType.toLowerCase()} occupying ${geography.toLowerCase()} terrain. Its capital, ${capitalName}, serves as the seat of power and the heart of its culture.${campaignContext ? ` In ${campaignContext}, it sits at the centre of the main conflict.` : ""}
-
-### Government & Power
-${rulerName} leads through a system of ${governmentStyle.toLowerCase()}. The realm's magic level is ${magicLevel.toLowerCase()}, which shapes both its military capability and its social order. ${trouble}.
-
-### Society & Culture
-The population lives under the tension of a realm at ${conflictLevel.toLowerCase()}. Loyalty is currency, and old alliances are being quietly renegotiated. The ruling class projects stability but its grip is tested daily.
-
-### How to use it at the table
-Use ${realmName} as the political backdrop for a campaign, the origin of a patron, or the prize in a succession conflict. The hidden trouble gives any visit the potential to escalate into something bigger.`;
-
-    const lore = `### At a Glance
-- **Polity Type**: ${polityType}
-- **Ruler**: ${rulerName} — calculating, cautious, and sitting on a secret
-- **Capital**: ${capitalName} — walled, busy, and watched
-- **Scale**: ${scale}
-- **Magic Level**: ${magicLevel}
-- **Conflict Level**: ${conflictLevel}
-- **Hidden Problem**: ${trouble}
-- **Immediate Hook**: A royal messenger arrived three days ago and has not been seen since
-
-### Major Factions
-- **${factionName1}**: Controls access to the realm's key resource and wants more influence at court
-- **${factionName2}**: Opposes the current ruler — quietly, for now
-
-### Rumours & Hooks
-- The ruler has been sending unmarked correspondence to a foreign power
-- A border garrison was found abandoned — no bodies, no signs of battle
-- ${factionName1} is recruiting outside its usual membership
-
-### Entity Seeds
-- **Character**: ${rulerName} (ruler)
-- **Faction**: ${factionName1}
-- **Faction**: ${factionName2}
-- **Location**: ${capitalName} (capital)
-- **Event**: The hidden crisis behind ${trouble.split("—")[0].trim()}`;
-
-    return {
-      type: "faction",
-      title: realmName,
-      summary,
-      content,
-      lore,
-      labels: ["rpg-kingdom", "kingdom-generator", "imported-draft"],
-      status: "active",
-    };
+    return generateKingdom(this.clientManager, options);
   }
 
   async generateNation(
-    options: {
-      genre?: string;
-      polityType?: string;
-      governmentStyle?: string;
-      scale?: string;
-      conflictLevel?: string;
-      campaignContext?: string;
-      useAI?: boolean;
-    } = {},
+    options: Parameters<typeof generateNation>[1] = {},
   ): Promise<GeneratorOutput> {
-    const genre =
-      options.genre ||
-      nationConfig.genres[
-        Math.floor(Math.random() * nationConfig.genres.length)
-      ];
-    const polityTypes =
-      nationConfig.polityTypesByGenre[genre] ??
-      nationConfig.polityTypesByGenre["Fantasy"];
-    const polityType =
-      options.polityType ||
-      polityTypes[Math.floor(Math.random() * polityTypes.length)];
-    const governmentStyle =
-      options.governmentStyle ||
-      nationConfig.governmentStyles[
-        Math.floor(Math.random() * nationConfig.governmentStyles.length)
-      ];
-    const scale = options.scale || nationConfig.scales[2];
-    const conflictLevel =
-      options.conflictLevel ||
-      nationConfig.conflictLevels[
-        Math.floor(Math.random() * nationConfig.conflictLevels.length)
-      ];
-    const campaignContext = options.campaignContext?.trim();
-    const varianceSeed = Math.floor(Math.random() * 99991) + 10;
-
-    if (options.useAI !== false) {
-      try {
-        const systemInstruction = `You are an expert RPG campaign writer. You generate immediately usable political entities for tabletop GMs in JSON format. You write in the register of the specified genre — a cyberpunk megacorp-state reads nothing like a fantasy kingdom.
-
-OUTPUT FORMAT — return ONLY a valid JSON object, no markdown fences:
-{
-  "title": "The entity's name — genre-appropriate and invented",
-  "summary": "One sentence: the entity's defining character and why a GM should use it.",
-  "content": "Markdown. Use exactly these four section headers in order: '### The State', '### Power Structure', '### Society & Tensions', '### How to use it at the table'. Each section: 3-5 tight sentences. Write in the genre's register. Include campaign context if provided.",
-  "lore": "Markdown. Use EXACTLY this structure:\\n### At a Glance\\n- **Type**: polity type\\n- **Leader / Authority**: name and one-line description (genre-appropriate)\\n- **Centre of Power**: name and one-line description\\n- **Scale**: territory/population summary\\n- **Conflict Level**: current state\\n- **Hidden Problem**: the tension simmering beneath the surface\\n- **Immediate Hook**: one-sentence GM hook\\n### Power Blocs\\n- **Name**: one-line description (2-3 factions/blocs)\\n### Rumours & Hooks\\n- short hook (2-3 bullet points)\\n### Entity Seeds\\n- list of 4-5 Codex entity types that naturally emerge from this state",
-  "labels": ["2-4 lowercase genre-appropriate tags, plus 'rpg-nation', 'nation-generator', 'imported-draft'"]
-}
-
-QUALITY RULES:
-- Everything — names, concerns, technology, terminology — must fit the genre.
-- The leader/authority must have a name, a defining trait, and a secret motive.
-- Each power bloc should have an agenda that creates tension.
-- Rumours should be specific and actionable.
-- Entity seeds should suggest concrete Codex entries a GM would actually create.`;
-
-        const userMessage = `Generate a political entity. Variation seed: ${varianceSeed}.
-- Genre / Setting: ${genre}
-- Polity Type: ${polityType}
-- Government Style: ${governmentStyle}
-- Scale: ${scale}
-- Conflict Level: ${conflictLevel}${campaignContext ? `\n- Campaign Context: ${campaignContext}` : ""}`;
-
-        const model = await this.clientManager.getModel(
-          "",
-          "gemini-3.1-flash-lite",
-          systemInstruction,
-        );
-        const response = await model.generateContent(userMessage);
-        const text = response.response.text().trim();
-        const cleanText = text
-          .replace(/^```json\s*/i, "")
-          .replace(/```$/, "")
-          .trim();
-        const data = JSON.parse(cleanText);
-
-        return {
-          type: "faction",
-          title: data.title || "The Unnamed State",
-          summary: data.summary || "",
-          content: data.content || "",
-          lore: data.lore || "",
-          labels: Array.isArray(data.labels)
-            ? data.labels
-            : ["rpg-nation", "nation-generator", "imported-draft"],
-          status: "active",
-        };
-      } catch (err) {
-        console.warn(
-          "AI generation failed, falling back to local tables:",
-          err,
-        );
-      }
-    }
-
-    // Local fallback
-    const trouble =
-      nationConfig.troubles[
-        Math.floor(Math.random() * nationConfig.troubles.length)
-      ];
-    const leaderName = this.generateName();
-    const bloc1 = `The ${this.generateName()} ${["Bloc", "Council", "Front", "Syndicate", "Coalition"][Math.floor(Math.random() * 5)]}`;
-    const bloc2 = `The ${this.generateName()} ${["Faction", "Assembly", "Circle", "Committee", "Network"][Math.floor(Math.random() * 5)]}`;
-    const rawName = this.generateName().replace(/\s+The\s+\S+$/, "");
-    const stateName = `${rawName} ${polityType}`;
-    const capitalName = this.generateName().replace(/\s+The\s+\S+$/, "");
-
-    const summary = `A ${conflictLevel.toLowerCase()} ${polityType.toLowerCase()} operating under ${governmentStyle.toLowerCase()}.`;
-
-    const content = `### The State
-${stateName} is a ${scale.toLowerCase()} ${polityType.toLowerCase()} under ${governmentStyle.toLowerCase()}. Its centre of power, ${capitalName}, is where decisions are made and deals are struck.${campaignContext ? ` In ${campaignContext}, it is a key player in the main conflict.` : ""}
-
-### Power Structure
-${leaderName} holds authority, but that authority is contested. ${trouble}. Those closest to power know the cracks are showing.
-
-### Society & Tensions
-The population lives under a state at ${conflictLevel.toLowerCase()}. Loyalty is shifting, old alliances are being renegotiated, and the ruling structure is projecting strength it may not have.
-
-### How to use it at the table
-Use ${stateName} as a campaign backdrop, the source of a mission, or the prize in a power struggle. The hidden problem gives any encounter with its institutions the potential to spiral.`;
-
-    const lore = `### At a Glance
-- **Type**: ${polityType}
-- **Leader / Authority**: ${leaderName} — in control, for now
-- **Centre of Power**: ${capitalName}
-- **Scale**: ${scale}
-- **Conflict Level**: ${conflictLevel}
-- **Hidden Problem**: ${trouble}
-- **Immediate Hook**: A key official has gone silent — their office is stonewalling
-
-### Power Blocs
-- **${bloc1}**: Controls a critical resource or access point; wants formal recognition
-- **${bloc2}**: Opposes the current leadership — quietly building leverage
-
-### Rumours & Hooks
-- ${leaderName} has been meeting with representatives of a rival power in secret
-- A key installation went dark last week; the official explanation doesn't add up
-- ${bloc1} is moving money through channels it has no business using
-
-### Entity Seeds
-- **Character**: ${leaderName} (leader)
-- **Faction**: ${bloc1}
-- **Faction**: ${bloc2}
-- **Location**: ${capitalName} (centre of power)
-- **Event**: The crisis — ${trouble}`;
-
-    return {
-      type: "faction",
-      title: stateName,
-      summary,
-      content,
-      lore,
-      labels: ["rpg-nation", "nation-generator", "imported-draft"],
-      status: "active",
-    };
-  }
-
-  private getRandomItems<T>(arr: T[], count: number): T[] {
-    const shuffled = [...arr].sort(() => 0.5 - Math.random());
-    return shuffled.slice(0, count);
+    return generateNation(this.clientManager, options);
   }
 }
