+++ apps/web/src/lib/services/seo/generators/base.ts
@@ -1,0 +1,421 @@
+export const nameTable = {
+  prefixes: [
+    "Ael",
+    "Bran",
+    "Cael",
+    "Dax",
+    "El",
+    "Fael",
+    "Grom",
+    "Had",
+    "Ior",
+    "Jor",
+    "Kael",
+    "Lyr",
+    "Morg",
+    "Nal",
+    "Odh",
+    "Py",
+    "Quel",
+    "Rhu",
+    "Sari",
+    "Thor",
+    "Uth",
+    "Vael",
+    "Wulf",
+    "Xan",
+    "Ysv",
+    "Zeph",
+  ],
+  suffixes: [
+    "dar",
+    "gorn",
+    "thor",
+    "ius",
+    "eth",
+    "wen",
+    "dolf",
+    "ric",
+    "mar",
+    "mon",
+    "las",
+    "vian",
+    "dor",
+    "morn",
+    "ra",
+    "thas",
+    "val",
+    "gar",
+    "rin",
+    "kis",
+    "on",
+    "ar",
+    "en",
+    "is",
+    "a",
+    "os",
+  ],
+  descriptors: [
+    "The Brave",
+    "The Silent",
+    "The Swift",
+    "The Bold",
+    "The Seer",
+    "The Wise",
+    "The Shadow",
+    "The Iron",
+    "The Wild",
+    "The Grim",
+    "The Lost",
+    "The Seeker",
+    "The Wanderer",
+    "The Exile",
+    "The Fierce",
+    "The Chronicler",
+  ],
+};
+
+export const nameGeneratorConfig = {
+  cultures: [
+    "Generic Fantasy",
+    "High Elf",
+    "Dark Elf",
+    "Dwarven",
+    "Halfling",
+    "Orcish",
+    "Norse / Viking",
+    "Roman / Latin",
+    "Celtic / Gaelic",
+    "Eastern / Asian-inspired",
+  ],
+  genders: ["Any", "Masculine", "Feminine", "Neutral / Ambiguous"],
+  nameTypes: ["Person", "Place", "Faction", "Item"],
+  counts: ["3", "5", "10"],
+  culturePrefixes: {
+    "Generic Fantasy": [
+      "Ael",
+      "Bran",
+      "Cael",
+      "Dax",
+      "El",
+      "Fael",
+      "Kael",
+      "Lyr",
+      "Morg",
+      "Nal",
+      "Thor",
+      "Vael",
+    ],
+    "High Elf": [
+      "Aer",
+      "Caer",
+      "El",
+      "Gal",
+      "Ith",
+      "Lae",
+      "Mir",
+      "Sil",
+      "Thal",
+      "Var",
+      "Wyn",
+      "Zin",
+    ],
+    "Dark Elf": [
+      "Driz",
+      "Mal",
+      "Nath",
+      "Riz",
+      "Sol",
+      "Szor",
+      "Ul",
+      "Vir",
+      "Xan",
+      "Zaer",
+      "Zin",
+      "Vy",
+    ],
+    Dwarven: [
+      "Bal",
+      "Dur",
+      "Glor",
+      "Grim",
+      "Kat",
+      "Kaz",
+      "Mag",
+      "Nar",
+      "Thor",
+      "Tor",
+      "Ulf",
+      "Vor",
+    ],
+    Halfling: [
+      "Bil",
+      "Cor",
+      "Del",
+      "Fil",
+      "Gob",
+      "Mer",
+      "Pip",
+      "Rose",
+      "Sam",
+      "Tom",
+      "Wil",
+      "Bun",
+    ],
+    Orcish: [
+      "Druk",
+      "Grak",
+      "Grom",
+      "Krag",
+      "Mog",
+      "Nar",
+      "Rok",
+      "Skul",
+      "Thok",
+      "Urg",
+      "Vrak",
+      "Zug",
+    ],
+    "Norse / Viking": [
+      "Arn",
+      "Bjor",
+      "Dag",
+      "Eil",
+      "Gunn",
+      "Hal",
+      "Ing",
+      "Leif",
+      "Rag",
+      "Sigr",
+      "Thor",
+      "Ulf",
+    ],
+    "Roman / Latin": [
+      "Aem",
+      "Aur",
+      "Bru",
+      "Cal",
+      "Cas",
+      "Cor",
+      "Flav",
+      "Jul",
+      "Marc",
+      "Octa",
+      "Sext",
+      "Val",
+    ],
+    "Celtic / Gaelic": [
+      "Aed",
+      "Bran",
+      "Caill",
+      "Conn",
+      "Donn",
+      "Eoch",
+      "Fearg",
+      "Niall",
+      "Ruad",
+      "Taig",
+      "Uar",
+      "Cunn",
+    ],
+    "Eastern / Asian-inspired": [
+      "Akira",
+      "Chen",
+      "Hiro",
+      "Jin",
+      "Kai",
+      "Li",
+      "Ren",
+      "Ryu",
+      "Shen",
+      "Tao",
+      "Wei",
+      "Zhen",
+    ],
+  } as Record<string, string[]>,
+  cultureSuffixes: {
+    "Generic Fantasy": [
+      "dar",
+      "eth",
+      "gorn",
+      "ius",
+      "mar",
+      "morn",
+      "ra",
+      "ric",
+      "thas",
+      "val",
+      "wen",
+      "thor",
+    ],
+    "High Elf": [
+      "ael",
+      "aith",
+      "ara",
+      "ath",
+      "elan",
+      "iath",
+      "ien",
+      "ira",
+      "is",
+      "ith",
+      "on",
+      "rial",
+    ],
+    "Dark Elf": [
+      "ace",
+      "arn",
+      "bra",
+      "dra",
+      "fein",
+      "ica",
+      "inae",
+      "ra",
+      "rix",
+      "ssa",
+      "tra",
+      "vra",
+    ],
+    Dwarven: [
+      "ak",
+      "ald",
+      "ar",
+      "ek",
+      "im",
+      "in",
+      "kin",
+      "nar",
+      "rek",
+      "rim",
+      "rin",
+      "uk",
+    ],
+    Halfling: [
+      "ble",
+      "boro",
+      "buck",
+      "fast",
+      "foot",
+      "fur",
+      "good",
+      "hill",
+      "penny",
+      "seed",
+      "wise",
+      "wool",
+    ],
+    Orcish: [
+      "ash",
+      "dar",
+      "gash",
+      "grak",
+      "gruk",
+      "kul",
+      "lash",
+      "mok",
+      "rak",
+      "skrag",
+      "thak",
+      "ugh",
+    ],
+    "Norse / Viking": [
+      "ald",
+      "bjorn",
+      "dís",
+      "eit",
+      "gar",
+      "helm",
+      "nar",
+      "rún",
+      "stein",
+      "ulf",
+      "vald",
+      "var",
+    ],
+    "Roman / Latin": [
+      "anus",
+      "ella",
+      "ia",
+      "ianus",
+      "illa",
+      "ina",
+      "inus",
+      "io",
+      "ius",
+      "lia",
+      "ona",
+      "us",
+    ],
+    "Celtic / Gaelic": [
+      "ach",
+      "all",
+      "an",
+      "ard",
+      "as",
+      "dha",
+      "ech",
+      "en",
+      "inn",
+      "on",
+      "ua",
+      "ugh",
+    ],
+    "Eastern / Asian-inspired": [
+      "bo",
+      "fa",
+      "ji",
+      "ko",
+      "lan",
+      "lei",
+      "li",
+      "mei",
+      "na",
+      "ro",
+      "shi",
+      "xia",
+    ],
+  } as Record<string, string[]>,
+};
+
+export interface GeneratorOutput {
+  type:
+    | "character"
+    | "creature"
+    | "location"
+    | "item"
+    | "event"
+    | "faction"
+    | "note";
+  title: string;
+  summary?: string;
+  content: string;
+  lore: string;
+  labels: string[];
+  status: "active" | "draft";
+}
+
+export function pickFrom<T>(arr: T[]): T {
+  return arr[Math.floor(Math.random() * arr.length)];
+}
+
+export function getRandomItems<T>(arr: T[], count: number): T[] {
+  const shuffled = [...arr].sort(() => 0.5 - Math.random());
+  return shuffled.slice(0, count);
+}
+
+export function generateName(): string {
+  const prefix =
+    nameTable.prefixes[Math.floor(Math.random() * nameTable.prefixes.length)];
+  const suffix =
+    nameTable.suffixes[Math.floor(Math.random() * nameTable.suffixes.length)];
+  const useDescriptor = Math.random() > 0.6;
+  let name = `${prefix}${suffix}`;
+  if (useDescriptor) {
+    const descriptor =
+      nameTable.descriptors[
+        Math.floor(Math.random() * nameTable.descriptors.length)
+      ];
+    name = `${name} ${descriptor}`;
+  }
+  return name;
+}
